% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAdjustGrassi2021.R
\name{calcAdjustGrassi2021}
\alias{calcAdjustGrassi2021}
\title{calcAdjustGrassi2021}
\usage{
calcAdjustGrassi2021()
}
\value{
magpie object with emission adjustement factors weighted by country C removals 2000 to 2015.
}
\description{
Calculates the carbon emission adjustement factors as derived by Grassi et al. 2021
Adjustement factors are provided in GtCO2 yr-1. Positive and negative values possible.
}
\examples{

\dontrun{
calcOutput("AdjustGrassi2021")
}
}
\author{
Michael Windisch, Florian Humpenoeder
}
