% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAgCapLabourShare.R
\name{calcAgCapLabourShare}
\alias{calcAgCapLabourShare}
\title{calcAgCapLabourShare}
\usage{
calcAgCapLabourShare()
}
\value{
MAgPIE object
}
\description{
This function calculates historical capital shares (Capital + Labour) of the factor requirements using USDA
}
\examples{
\dontrun{ 
calcOutput("calcAgCapLabourShare")
}

}
\seealso{
\code{\link{calcAgCapLabourShare}}
}
\author{
Edna J. Molina Bacca
}
