% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcAgEmplILO.R
\name{calcAgEmplILO}
\alias{calcAgEmplILO}
\title{calcAgEmplILO}
\usage{
calcAgEmplILO()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates complete dataset of number of people employed in agriculture, forestry and fishery based
on ILO modelles dataset and GDPpcPPP05 for regression
}
\examples{
\dontrun{
calcOutput("AgEmplILO")
}
}
\author{
Debbora Leip
}
