% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBEYield.R
\name{calcBEYield}
\alias{calcBEYield}
\title{calcBEYield}
\usage{
calcBEYield(returnWeights = FALSE)
}
\arguments{
\item{returnWeights}{if TRUE returns crop-specific cellular cropland area weights
(begr/betr) instead of the yields}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Calculates bioenergy crop yields from Li et al. (2020) for the two
MAgPIE bioenergy types begr and betr. begr takes the maximum yield across
Miscanthus and Switchgrass per cell. betr takes the maximum yield across
Eucalypt (warm/tropical climates only), Poplar, and Willow per cell.
}
\examples{
\dontrun{
calcOutput("BEYield", aggregate = FALSE)
calcOutput("BEYield", returnWeights = TRUE, aggregate = FALSE)
}
}
\seealso{
\code{\link[=readLi2020]{readLi2020()}}
}
\author{
Kristine Karstens
}
