% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiocharAttributes.R
\name{calcBiocharAttributes}
\alias{calcBiocharAttributes}
\title{calcBiocharAttributes}
\usage{
calcBiocharAttributes(processCond = "default")
}
\arguments{
\item{processCond}{Selects the pyrolysis process conditions for which data is returned.
Options: "all", "default", or subset of:
"SP-400", "SP-500", "SP-650", "SP-800", "FP-500", "P-Woolf", "P-Buffi", "P-KonTiki".}
}
\value{
List of magpie objects with results on global level, empty weight,
unit and description.
}
\description{
Provides biochar attributes for different pyrolysis process conditions.
Currently covered are Dry Matter (DM), Carbon (C), and Generalizable Energy (GE).
Values are assembled from various literature sources.
}
\examples{
\dontrun{
calcOutput("BiocharAttributes", processCond = "all")
}
}
\seealso{
\code{\link[=readPyrolysisConditions]{readPyrolysisConditions()}}
}
\author{
Isabelle Weindl
}
