% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiocharEfficiency.R
\name{calcBiocharEfficiency}
\alias{calcBiocharEfficiency}
\title{calcBiocharEfficiency}
\usage{
calcBiocharEfficiency(processCond = "default", outputType = "default")
}
\arguments{
\item{processCond}{Selects the pyrolysis process conditions for which data is returned.
Options: "all", "default", or subset of:
"SP-400", "SP-500", "SP-650", "SP-800", "FP-500", "P-Woolf", "P-Buffi", "P-KonTiki".}

\item{outputType}{Selects the output metric for which data is returned.
Options: "all", "default", or subset of:
"dm_yield", "en_yield", "c_yield", "HC_ratio".}
}
\value{
List of magpie objects with results on global level, empty weight,
unit and description.
}
\description{
Calculates biochar conversion efficiencies for different pyrolysis
process conditions, using data from multiple literature sources.
Yield metrics are expressed as output biochar per biomass input on dry matter,
carbon, and energy basis. Output can also cover H/C ratio when available.
}
\examples{
\dontrun{
calcOutput("BiocharEfficiency", processCond = "all", outputType = "all")
}
}
\seealso{
\code{\link[=readPyrolysisConditions]{readPyrolysisConditions()}}
}
\author{
Isabelle Weindl
}
