% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiomassForNonFood.R
\name{calcBiomassForNonFood}
\alias{calcBiomassForNonFood}
\title{calcBiomassForNonFood}
\usage{
calcBiomassForNonFood()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the supply of agricultural and forestry based biomass
that are used in the energy or industry sector
}
\examples{
\dontrun{
calcOutput("calcBiomassForNonFood")
}
}
\author{
Kristine Karstens
}
