% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBiomeType.R
\name{calcBiomeType}
\alias{calcBiomeType}
\title{calcBiomeType}
\usage{
calcBiomeType(cells = "magpiecell")
}
\arguments{
\item{cells}{magpiecell (59199 cells) or lpjcell (67420 cells)}
}
\value{
List with a magpie object
}
\description{
Returns fraction of spatial unit (cell) belonging to a biome type of each biogeographic realm. The classification is based on data from 'the nature conservancy' (https://geospatial.tnc.org/datasets/b1636d640ede4d6ca8f5e369f2dc368b/about) originally developed by Olson et al. (2001), BioScience.
}
\examples{
\dontrun{
calcOutput("BiomeType", aggregate = FALSE)
}

}
\seealso{
\code{\link{readTNC2019}}
}
\author{
Patrick v. Jeetze
}
