% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcBioplasticToBiomass.R
\name{calcBioplsticToBiomass}
\alias{calcBioplsticToBiomass}
\title{calcBioplsticToBiomass}
\usage{
calcBioplsticToBiomass()
}
\value{
List of magpie objects with global conversion factors, unit and description.
}
\description{
calculates conversion factors from bioplastic demand to demand of biomass needed for the
production, taking into account the average share of different biomass sources (glycerol, starch, sugars, cellulose,
oils) for bioplastic production and corresponding content in the different crop types
}
\examples{
\dontrun{
calcOutput("BioplsticToBiomass")
}
}
\author{
Debbora Leip
}
