% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcClossConfinement.R
\name{calcClossConfinement}
\alias{calcClossConfinement}
\title{calcClossConfinement}
\usage{
calcClossConfinement()
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Carbon losses for livestock confinements
}
\examples{
\dontrun{
calcOutput("ClossConfinement")
}
}
\author{
Kristine Karstens
}
