% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcConstructionWoodDemand.R
\name{calcConstructionWoodDemand}
\alias{calcConstructionWoodDemand}
\title{calcConstructionWoodDemand}
\usage{
calcConstructionWoodDemand()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the demand of construction wood from Galina et al. 2020 data.
See
}
\examples{
\dontrun{
calcOutput("ConstructionWoodDemand")
}
}
\seealso{
\code{\link[mrfaocore:calcFAOmassbalance_pre]{mrfaocore::calcFAOmassbalance_pre()}}
}
\author{
Abhijeet Mishra
}
