% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcDevelopmentState.R
\name{calcDevelopmentState}
\alias{calcDevelopmentState}
\title{calc}
\usage{
calcDevelopmentState(upper = 12746, lower = 1045)
}
\arguments{
\item{upper}{Change upper limit (default: 12746, i.e. the upper threshold for 2013 in current US$)}

\item{lower}{Change lower limit (default: 1045, i.e. the lower threshold for 2013 in current US$)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Provides development state of a country or region.
We use worldbank definitions by default: above 12746 USD per capita, its a
high-income country, below 1045 its a low-income country, in between its a
medium-income country.
}
\examples{
\dontrun{
calcOutput("DevelopmentState")
}

}
\seealso{
\code{\link[mrdrivers:calcGDP]{mrdrivers::calcGDPpc()}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
