% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEDGAR6.R
\name{calcEDGAR6}
\alias{calcEDGAR6}
\title{calcEDGAR6}
\usage{
calcEDGAR6(non_country = FALSE, nutrient = TRUE)
}
\arguments{
\item{non_country}{can alternatively provide SEA and AIR emissions, which are global emissions that
cannot be attributed to a country}

\item{nutrient}{if TRUE, nitrogen is reported as N and CO2 as C}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Creates an emission inventory based on the EDGAR5 and EDGAR6 database.
}
\author{
Benjamin Leon Bodirsky
}
