% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcEmiMacCEDS.R
\name{calcEmiMacCEDS}
\alias{calcEmiMacCEDS}
\title{Calculate baseline emissions for maccs (mostly) from CEDS2024 data}
\usage{
calcEmiMacCEDS(baseyear = 2020)
}
\arguments{
\item{baseyear}{year to take as a reference from CEDS, ignored for the EDGAR2005 LUC CO2 emissions}
}
\value{
REMIND data for baseline emissions for maccs and corresonding
weights (NULL) as a list of two MAgPIE objects
}
\description{
Provides REMIND data for baseline emissions for maccs from CEDS2024
Note that this can't fully replace the old EDGAR data, as it doesn't include
land use CO2 emissions. These are not currently actually used in REMIND
by default as the LUC CO2 MACC is off, but it's still not officially deprecated.
}
\details{
Therefore, the CO2 from LUC is read from EDGAR with a 2005, a potentially
different baseyear than the one chosen in the parameter
}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}, \code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Gabriel Abrahao
}
