% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFacReq.R
\name{calcFacReq}
\alias{calcFacReq}
\title{calcFacReq}
\usage{
calcFacReq(splitSectors = FALSE)
}
\arguments{
\item{splitSectors}{if TRUE requirements for labor and capital will be reported separately}
}
\value{
MAgPIE object
}
\description{
This function calculates factor requirement costs (05USDMER/tDM) on regional level,
based FAO databases
}
\examples{
\dontrun{
calcOutput("calcFacReq")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}},\code{\link[=calcFactorIntensity]{calcFactorIntensity()}}
}
\author{
Debbora Leip
}
