% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFactorCostsCrops.R
\name{calcFactorCostsCrops}
\alias{calcFactorCostsCrops}
\title{calcFactorCostsCrops}
\usage{
calcFactorCostsCrops(datasource = "USDA")
}
\arguments{
\item{datasource}{only source available is "USDA" (calculates factor costs by applying factor cost share from USDA
to VoP from FAO)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates factor costs for crop production in mio. US$MER05
}
\examples{
\dontrun{
calcOutput("FactorCostsCrops")
}
}
\author{
Debbora Leip
}
