% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFactorIntensity.R
\name{calcFactorIntensity}
\alias{calcFactorIntensity}
\title{calcFactorIntensity}
\usage{
calcFactorIntensity(output = "intensities", method = "USDA")
}
\arguments{
\item{output}{needed outputs. It can be either "intensities" (Capital/Labour factor intensities),
"requirements" (Capital Stock requirements per ton), and "CapitalShare" for "USDA" method.
For the "CapitalStock" method only "intensities" and "requirements" outputs supported.}

\item{method}{"USDA" or "CapitalStock"}
}
\value{
magpie object of the factor requirements intensity or factor intensity in 05USDppp/tDM per crop,
or capital share fraction.
}
\description{
Calculates factor intensity for labour and/or capital from USDA (Inputs share)
and FAO (Value of Production)in 05USDppp per ton.
Capital intensity and requirements can also be calculated from FAO's CapitalStock database.
}
\examples{
\dontrun{
a <- calcOutput("FactorIntensity")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Edna J. Molina Bacca
}
