% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFallowLandOLD.R
\name{calcFallowLandOLD}
\alias{calcFallowLandOLD}
\title{calcFallowLandOLD}
\usage{
calcFallowLandOLD()
}
\value{
Magpie object with fallow land in ha
}
\description{
Calculates fallow land on grid cell level,
based on physical cropland extend and harvested area output
of LPJML io toolbox.
The formula
"fallow land are = max( physical cropland area - harvested cropland area, 0)"
is used.
Due to multiple cropping, harvested cropland area can be greater than non-fallow land area
and even greater than physical cropland area.
Thus the results can only be considered a rough estimate of fallow land area.
}
\examples{
\dontrun{
calcOutput("FallowLandOLD")
}
}
\seealso{
\code{\link{readLanduseToolbox}}
}
\author{
David Hoetten
}
