% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOBilateralTrade.R
\name{calcFAOBilateralTrade}
\alias{calcFAOBilateralTrade}
\title{calcFAOBilateralTrade}
\usage{
calcFAOBilateralTrade(
  output = "value",
  products = "kcr",
  prod_agg = TRUE,
  five_year = TRUE
)
}
\arguments{
\item{output}{"value", "qty", or "price"}

\item{products}{"kcr", "kli", or "kothers"}

\item{prod_agg}{binary to keep FAO product level or magpie}

\item{five_year}{only 5 year steps due to memory load}
}
\value{
List of magpie objects with results on bilateral country level, weight on bilateral country level, unit and description.
}
\description{
Calculates bilateral trade values based on FAO trade matrix
}
\examples{

\dontrun{
calcOutput("BilateralTradeValue")
}

}
\author{
David M Chen
}
