% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOIntraYearProd.R
\name{calcFAOIntraYearProd}
\alias{calcFAOIntraYearProd}
\title{calcFAOIntraYearProd}
\usage{
calcFAOIntraYearProd(
  day = "harvest_day",
  products = "kcr",
  frequency = "monthly",
  attribute = "dm"
)
}
\arguments{
\item{day}{harvest_day (to market) or maturity_day (first mature)}

\item{products}{"kcr" or "staples" staples uses FAO production dataset instead of calcProduction
to only give maize wheat soy and rice. Allows for more years. A bit of a David-specific subtype}

\item{frequency}{monthly or quarterly. Daily leads to memory limits.}

\item{attribute}{dm default. can only select one at a time due to memory

#' @seealso
\code{\link{readGGCMICropCalendar}}}
}
\description{
Distribute massbalanced or FAOSTAT staple production to monthly or quarterly interval based on GGCMI crop calendar.
Only national level implemented for now as cellular production only available on 5 year time steps due to memory.
Assume rainfed crop calendar date for now.
}
\author{
David Chen
}
