% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOmassbalance_pre.R
\name{calcFAOmassbalance_pre}
\alias{calcFAOmassbalance_pre}
\title{calcFAOmassbalance_pre}
\usage{
calcFAOmassbalance_pre(years = NULL)
}
\arguments{
\item{years}{years to be estimated, if null, then all years in FAOharmonized are returned}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
This is an intermediary result, which is used e.g. for estimating the feed baskets. For most uses, it is more appropriate to use the FAOmasbalance instead of the FAOmassbalance_pre.
}
\description{
Calculates an extended version of the Food Balance Sheets. Makes explicit the conversion processes that convert one type of product into another.
Includes processes like milling, distilling, extraction etc. Adds certain byproducts like distillers grains or ethanol.
}
\examples{
\dontrun{
calcOutput("FAOmassbalance_pre")
}
}
\seealso{
\code{\link[=calcFAOmassbalance]{calcFAOmassbalance()}}
}
\author{
Benjamin Leon Bodirsky
}
