% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOTradePrices.R
\name{calcFAOTradePrices}
\alias{calcFAOTradePrices}
\title{calcFAOTradePrices}
\usage{
calcFAOTradePrices(aggregation = "k")
}
\arguments{
\item{aggregation}{"none", "k", "fbs" or "springmann" the last uses Marco Springmann's custom product mapping}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates USD per kg of FAOSTAT Trade data for import and export prices
}
\examples{
\dontrun{
calcOutput("calcFAOTradePrices")
}

}
\author{
David M Chen
}
