% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOYield.R
\name{calcFAOYield}
\alias{calcFAOYield}
\title{calcFAOYield}
\usage{
calcFAOYield(
  physical = TRUE,
  attributes = "dm",
  irrigation = FALSE,
  cellular = FALSE,
  cut = FALSE,
  average = 5,
  areaSource = "LandInG"
)
}
\arguments{
\item{physical}{physical area or havested area}

\item{attributes}{in dm, wm, ge, nr, p, k}

\item{irrigation}{distinguish irrigation or not}

\item{cellular}{if TRUE value is calculate on cellular level}

\item{cut}{FALSE (default) - do not cut off yields,
number between 0 and 1 to define percentile value for cut off}

\item{average}{averaging period in years (if NULL no averaging is used)}

\item{areaSource}{data source for croparea used in calculation: "LandInG" (default)
or "FAOLUH". The latter one is deprecated. The former default
"FAO" is no longer accepted.}
}
\value{
MAgPIE object of yields
}
\description{
calculates the yield based on FAO data
}
\author{
Debbora Leip, Jan Philipp Dietrich, Kristine Karstens, Felicitas Beier
}
