% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBalanceflow.R
\name{calcFeedBalanceflow}
\alias{calcFeedBalanceflow}
\title{calcFeedBalanceflow}
\usage{
calcFeedBalanceflow(
  per_livestock_unit = FALSE,
  cellular = FALSE,
  cells = "lpjcell",
  products = "kall",
  future = "constant"
)
}
\arguments{
\item{per_livestock_unit}{default false}

\item{cellular}{if TRUE value is calculated on cellular level}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420)}

\item{products}{products in feed baskets that shall be reported}

\item{future}{if FALSE, only past years will be reported (reduces memory)}
}
\value{
List of magpie objects with results on country or cellular level, unit and description.
}
\description{
Calculates feed balance flows from MAgPIE-Feed model to meet FAO data
}
\examples{
\dontrun{
calcOutput("FeedBalanceflow")
}
}
\author{
Isabelle Weindl, Kristine Karstens
}
