% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBasketsUncalibrated.R
\name{calcFeedBasketsUncalibrated}
\alias{calcFeedBasketsUncalibrated}
\title{calcFeedBasketsUncalibrated}
\usage{
calcFeedBasketsUncalibrated()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Combines uncalibrated feed baskets of the past with scenario-dependent future feed baskets.
}
\examples{
\dontrun{
calcOutput("FeedBasketsUncalibrated")
}
}
\author{
Isabelle Weindl, Benjamin Leon Bodirsky, Stephen Wirth, Jan Philipp Dietrich
}
