% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFertilizerPricesFAO.R
\name{calcFertilizerPricesFAO}
\alias{calcFertilizerPricesFAO}
\title{calcFertilizerPricesFAO}
\usage{
calcFertilizerPricesFAO(subtype = "N", by = "nutrient")
}
\arguments{
\item{subtype}{"N" for fertilizer containing nitrogen, "P" for fertilizer containing phosphorus}

\item{by}{"nutrient" if referring to price per amount of nutrients (N or P) within the fertilizer products, or
"product" if referring to price per amount of fertilizer product}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates dataset of fertilizer prices in US$MER05/tonne (either referring to the amount of fertilizer
product, or to the amount of nutrients within the fertilizer) based on FAO data
}
\examples{
\dontrun{
calcOutput("FertilizerPricesFAO", subtype = "N", by = "nutrient")
}
}
\author{
Debbora Leip
}
