% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFertilizerUseFAO.R
\name{calcFertilizerUseFAO}
\alias{calcFertilizerUseFAO}
\title{calcFertilizerUseFAO}
\usage{
calcFertilizerUseFAO(subtype = "N", by = "nutrient")
}
\arguments{
\item{subtype}{"N" for fertilizer containing nitrogen, "P" for fertilizer containing phosphorus (note that there
is an overlap between those categories, as some fertilizers include both nutrients)}

\item{by}{"nutrient" if referring to amount of nutrients (N or P) in total used fertilizer, or "product" if
referring to total amount of fertilizer used}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates dataset of fertilizer use in tonnes (either referring to the amount of fertilizer products
used, or to the amount of nutrients within the fertilizer used) based on FAO data
}
\examples{
\dontrun{
calcOutput("FertilizerUseFAO", subtype = "N", by = "nutrient")
}

}
\author{
Debbora Leip
}
