% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFoodWasteAndSewage.R
\name{calcFoodWasteAndSewage}
\alias{calcFoodWasteAndSewage}
\title{calcFoodWasteAndSewage}
\usage{
calcFoodWasteAndSewage(historic = TRUE)
}
\arguments{
\item{historic}{If set to TRUE only ho}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates food waste and sewage based on the FAO mass balance or future
projections from Bodirsky et al. (2018).
}
\examples{
\dontrun{
calcOutput("FoodWasteAndSewage")
}
}
\seealso{
\code{\link[=calcFAOmassbalance]{calcFAOmassbalance()}}
}
\author{
Benjamin Leon Bodirsky, Xiaoxi Wang
}
