% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFractionInputsUSDA.R
\name{calcFractionInputsUSDA}
\alias{calcFractionInputsUSDA}
\title{calcFractionInputsUSDA}
\usage{
calcFractionInputsUSDA(products = "kcr")
}
\arguments{
\item{products}{either "kcr" for crops, or "kli" for livestock}
}
\value{
magpie object of the shares of the factor requirements in agriculture (capital, labor, revenue, materials, land).
}
\description{
Calculates the factor factor shares for crop production from USDA'S Inputs shares.
}
\examples{
\dontrun{
a <- calcOutput("FractionInputsUSDA")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Edna J. Molina Bacca, Debbora Leip
}
