% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPppp.R
\name{calcGDPpppFuture}
\alias{calcGDPpppFuture}
\title{calcGDPpppFuture}
\usage{
calcGDPpppFuture(GDPpppFuture = NULL)
}
\arguments{
\item{GDPpppFuture}{GDPppp future data source}
}
\value{
GDP PPP(ICP11) in million USD05 equivalents
}
\description{
\verb{r lifecycle::badge('deprecated')}

This function is defunct. Use \code{\link[mrdrivers:calcGDPFuture]{mrdrivers::calcGDPFuture()}} instead.
}
\seealso{
\code{\link[mrdrivers:calcGDPFuture]{mrdrivers::calcGDPFuture()}}
}
