% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGovernanceIndicator.R
\name{calcGovernanceIndicator}
\alias{calcGovernanceIndicator}
\title{calcGovernanceIndicator}
\usage{
calcGovernanceIndicator()
}
\value{
magpie object at iso-country level
}
\description{
returns governance indicator from Andrijevic et al. 2019
}
\examples{
\dontrun{
calcOutput("GovernanceIndicator", aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
