% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingPeriodMonths.R
\name{calcGrowingPeriodMonths}
\alias{calcGrowingPeriodMonths}
\title{calcGrowingPeriodMonths}
\usage{
calcGrowingPeriodMonths(selectyears, lpjml, climatetype, minThreshold = 100)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{minThreshold}{Threshold of monthly grass GPP to be classified as
growing period month
Unit of the threshold is gC/m^2.
Default: 100gC/m^2
Note: the default value is chosen based on LPJmL version 5
to reflect multiple cropping suitability as shown in GAEZ-4.
An update of LPJmL5 with regards to grass management may
require an adjustment of the threshold.}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates which gridcell-specific months in which
growing conditions are favorable for crop growth
based on monthly grass GPP
}
\examples{
\dontrun{
calcOutput("GrowingPeriodMonths", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
