% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStock.R
\name{calcGrowingStock}
\alias{calcGrowingStock}
\title{calcGrowingStock}
\usage{
calcGrowingStock()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks on FAO data.
}
\examples{

\dontrun{ 
calcOutput("GrowingStock",aggregate=TRUE)
}

}
\author{
Abhijeet Mishra
}
