% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockNatVegAbsolute.R
\name{calcGrowingStockNatVegAbsolute}
\alias{calcGrowingStockNatVegAbsolute}
\title{calcGrowingStockNatVegAbsolute}
\usage{
calcGrowingStockNatVegAbsolute()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks on FAO data.
}
\examples{
\dontrun{
calcOutput("GrowingStockNatVegAbsolute", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra
}
