% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockNRF.R
\name{calcGrowingStockNRF}
\alias{calcGrowingStockNRF}
\title{calcGrowingStockNRF}
\usage{
calcGrowingStockNRF(indicator = "per_ha")
}
\arguments{
\item{indicator}{Either "per_ha" (m3/ha, area-weighted) or "total" (Mm3, absolute).}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks from FAO data for naturally regenerating forests
i.e. primary forests and secondary forests.
}
\examples{
\dontrun{
calcOutput("GrowingStockNRF", indicator = "per_ha", aggregate = TRUE)
calcOutput("GrowingStockNRF", indicator = "total", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra, Florian Humpenoeder
}
