% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockpha.R
\name{calcGrowingStockpha}
\alias{calcGrowingStockpha}
\title{calcGrowingStockpha}
\usage{
calcGrowingStockpha()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks on FAO data.
}
\examples{
\dontrun{
calcOutput("GrowingStockpha", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra
}
