% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockPlantAbsolute.R
\name{calcGrowingStockPlantAbsolute}
\alias{calcGrowingStockPlantAbsolute}
\title{calcGrowingStockPlantAbsolute}
\usage{
calcGrowingStockPlantAbsolute()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks on FAO data.
}
\examples{
\dontrun{
calcOutput("GrowingStockPlantAbsolute", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra
}
