% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockPlantations.R
\name{calcGrowingStockPlantations}
\alias{calcGrowingStockPlantations}
\title{calcGrowingStockPlantations}
\usage{
calcGrowingStockPlantations(indicator = "per_ha")
}
\arguments{
\item{indicator}{Either "per_ha" (m3/ha, area-weighted) or "total" (Mm3, absolute).}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks from FAO data for planted forests (plantations).
}
\examples{
\dontrun{
calcOutput("GrowingStockPlantations", indicator = "per_ha", aggregate = TRUE)
calcOutput("GrowingStockPlantations", indicator = "total", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra, Florian Humpenoeder
}
