% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGTAPTotalTransportCosts.R
\name{calcGTAPTotalTransportCosts}
\alias{calcGTAPTotalTransportCosts}
\title{calcGTAPTotalTransportCosts}
\usage{
calcGTAPTotalTransportCosts(costType = "transport", version = "9")
}
\arguments{
\item{costType}{transport or wholesale}

\item{version}{"81" or "9"}
}
\value{
List of magpie object with results on country level, unit and description.
}
\description{
Calculates iso-level total global transport costs from GTAP for GTAP food commodities
}
\examples{
\dontrun{
calcOutput("GTAPTotalTransportCosts")
}

}
\author{
David M Chen
}
