% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHistBioplasticProd.R
\name{calcHistBioplasticProd}
\alias{calcHistBioplasticProd}
\title{calcHistBioplasticProd}
\usage{
calcHistBioplasticProd()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates historic bioplastic production based on data and linear extrapolation
}
\examples{
\dontrun{
calcOutput("HistBioplasticProd")
}
}
\author{
Debbora Leip
}
