% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHistEmissions.R
\name{calcHistEmissions}
\alias{calcHistEmissions}
\title{historical emissions per sector or mac}
\usage{
calcHistEmissions(subtype = "sector")
}
\arguments{
\item{subtype}{Either "sector" or "MAC"}
}
\value{
magpie object of historical emissions data
}
\description{
Provides historical emissions values per economic sector or per mac sector.
For now it only includes European data.
}
\examples{
\dontrun{
calcOutput("HistEmissions")
}
}
\author{
Renato Rodrigues
}
