% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHourlyLaborCosts.R
\name{calcHourlyLaborCosts}
\alias{calcHourlyLaborCosts}
\title{calcHourlyLaborCosts}
\usage{
calcHourlyLaborCosts(
  datasource = "USDA_FAO",
  sector = "agriculture",
  fillWithRegression = TRUE,
  calibYear = 2010,
  projection = FALSE
)
}
\arguments{
\item{datasource}{either raw data from "ILO" (agriculture+forestry+fishery) or data calculated based on total labor
costs from "USDA_FAO" (crop+livestock production)}

\item{sector}{should average hourly labor costs be reported ("agriculture"), or hourly labor costs specific to
either "crops" or "livestock" production. For ILO only the aggregate hourly labor costs are available.}

\item{fillWithRegression}{boolean: should missing values be filled based on a regression between ILO hourly labor
costs and GDPpcMER (calibrated to countries)}

\item{calibYear}{in case of fillWithRegression being TRUE, data after this year will be ignored and calculated using
the regression (calibrated for each year to calibYear, or the most recent year with data before calibYear)}

\item{projection}{either FALSE or SSP on which projections should be based. Only relevant if fillWithRegression is
TRUE.}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates dataset of hourly labor costs per employee in agriculture
}
\examples{
\dontrun{
calcOutput("HourlyLaborCosts")
}
}
\author{
Debbora Leip
}
