% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcHourlyLaborCostsILO.R
\name{calcHourlyLaborCostsILO}
\alias{calcHourlyLaborCostsILO}
\title{calcHourlyLaborCostsILO}
\usage{
calcHourlyLaborCostsILO()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates complete dataset of hourly labor costs per employee in agriculture, forestry and
fishery based on ILO dataset and GDPpcMER05 for regression
}
\examples{
\dontrun{
calcOutput("HourlyLaborCostsILO")
}
}
\author{
Debbora Leip
}
