% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIFPRIsubsidy.R
\name{calcIFPRIsubsidy}
\alias{calcIFPRIsubsidy}
\title{calcIFPRIsubsidy}
\usage{
calcIFPRIsubsidy(fillGaps = TRUE)
}
\arguments{
\item{fillGaps}{boolean, should gaps in the dataset be filled using interpolation?}
}
\value{
magpie object. in mio. USDMER05
}
\description{
Adds non-allocated subsidies to crop subsidies (as most subsidies are linked to land area), and
excludes NRP subsidies (as those are border measures, which are already reflected in ag. prices)
}
\examples{
\dontrun{
a <- calcOutput("IFPRIsubsidy")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Debbora Leip
}
