% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIOEdgeBuildings.R
\name{calcIOEdgeBuildings}
\alias{calcIOEdgeBuildings}
\title{calcIOEdgeBuildings}
\usage{
calcIOEdgeBuildings(subtype = c("output_EDGE", "output_EDGE_buildings"))
}
\arguments{
\item{subtype}{Data subtype. See default argument for possible values.}
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Computes IEA-based model data for different subtypes by use of raw IEA Energy Balances data
and a mapping that corresponds to the structure of "products" and "flows" of IEA.
}
\details{
Mapping structure example: IEA product ANTCOAL used for IEA flow TPATFUEL, contributes via REMIND technology
coaltr for generating sesofos from pecoal (REMIND names)
}
\examples{
\dontrun{
a <- calcOutput("IOEdgeBuildings", subtype = "output_EDGE_buildings")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Pascal Führlich, Anastasis Giannousakis
}
