% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIPCCefNSoil.R
\name{calcIPCCefNSoil}
\alias{calcIPCCefNSoil}
\title{calcIPCCefNSoil}
\usage{
calcIPCCefNSoil()
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Emission factors for croplands based on IPCC using a country-level leaching fraction
}
\examples{

\dontrun{ 
calcOutput("IPCCefNSoil")
}

}
\author{
Benjamin Leon Bodirsky
}
