% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcISIMIP3bYields.R
\name{calcISIMIP3bYields}
\alias{calcISIMIP3bYields}
\title{calcISIMIP3bYields}
\usage{
calcISIMIP3bYields(
  subtype = "yields:EPIC-IIASA:ukesm1-0-ll:ssp585:default:3b",
  smooth = TRUE,
  cells = "magpiecell"
)
}
\arguments{
\item{subtype}{subtype of yield based on readISIMIPoutputs, for crop yields}

\item{smooth}{smooth cells via spline}

\item{cells}{magpie or lpjcell}
}
\value{
magpie object in cellular resolution
}
\description{
reads and cleans up ISIMIP3b crop yield data
}
\examples{
\dontrun{
calcOutput("ISIMIP3bYields", aggregate = FALSE)
}

}
\author{
David Meng-Chuen Chen
}
