% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLaborCosts.R
\name{calcLaborCosts}
\alias{calcLaborCosts}
\title{calcLaborCosts}
\usage{
calcLaborCosts(
  datasource = "ILO",
  subsectors = TRUE,
  inclFish = FALSE,
  inclForest = FALSE,
  otherLivst = TRUE,
  gtapVar = "NVFA",
  addSubsidies = FALSE
)
}
\arguments{
\item{datasource}{data source on which the labor costs should be based. Available are ILO, USDA (which also uses data
on VoP from FAO), and GTAP}

\item{subsectors}{boolean: should output be aggregated or split into available subsectors (crops, livst, forestry,
fishery)}

\item{inclFish}{boolean: should fish labor costs be included?}

\item{inclForest}{boolean: should forestry labor costs be included (only available for ILO and GTAP)?}

\item{otherLivst}{boolean: should other_livst category (i.e. beeswax, wool, silkworms, and honey) be included
in livestock (only relevant for datasource USDA)?}

\item{gtapVar}{variable name to use from GTAP (only relevant if source is "GTAP")}

\item{addSubsidies}{boolean: should subsidy data (from IFPRI) should be added to VoP before applying USDA cost
shares (only relevant if datasource is "USDA")}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates total labor costs in mio. US$MER05 (coverage depends on source: crop, livestock and
fish production for USDA, and additionally forestry for GTAP and ILO)
}
\examples{
\dontrun{
calcOutput("LaborCosts", datasource = "ILO")
}
}
\author{
Debbora Leip
}
