% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandArea.R
\name{calcLandArea}
\alias{calcLandArea}
\title{calcLandArea}
\usage{
calcLandArea(cells = "magpiecell")
}
\arguments{
\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
magpie object in cellular resolution
}
\description{
total land area in a cell
}
\examples{
\dontrun{
calcOutput("LandArea", aggregate = FALSE)
}

}
\author{
Jan Philipp Dietrich, Felicitas Beier
}
