% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLandEmissions.R
\name{calcLandEmissions}
\alias{calcLandEmissions}
\title{calcLandEmissions}
\usage{
calcLandEmissions(datasource = "CEDS")
}
\arguments{
\item{datasource}{The Emission Inventory that shall be used. For futher information, best see function calcEmissionInventory. Options are e.g.  CEDS, combined_CEDS_IPCC (including own estimates where available), IPCC(own estimates), Nsurplus (own estimates)}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
Land emission data. This function was originally called calcValidEmissions and located in
mrvalidation.
}
\examples{

\dontrun{
calcOutput("LandEmissions")
}
}
\author{
Benjamin Leon Bodirsky, Michael S. Crawford
}
