% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLanduseInitialisationBase.R
\name{calcLanduseInitialisationBase}
\alias{calcLanduseInitialisationBase}
\title{calcLanduseInitialisationBase}
\usage{
calcLanduseInitialisationBase(cells = "lpjcell", selectyears = "past")
}
\arguments{
\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}

\item{selectyears}{Years to be computed (default on "past")}
}
\value{
Cellular landuse initialisation in its base configuration
}
\description{
Calculates the cellular MAgPIE landuse initialisation area. Data from FAO on forestry is used
to split the secondary forest pool of the LU2v2 dataset into forestry and secd_forest. This function
returns the data set in a basic configuration. Use \code{\link{calcLanduseInitialisation}} for
more settings.
}
\examples{
\dontrun{
calcOutput("LanduseInitialisationBase")
}
}
\author{
Jan Philipp Dietrich, Benjamin Leon Bodirsky, Kristine Karstens, Felcitas Beier, Patrick v. Jeetze
}
