% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivestockProductivity.R
\name{calcLivestockProductivity}
\alias{calcLivestockProductivity}
\title{Calculate Yields for Livestock}
\usage{
calcLivestockProductivity(future = TRUE)
}
\arguments{
\item{future}{if TRUE calculates constant future and linear trends based on SSP Expert guesses}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Provides livestock yields for historical and future years.
}
\examples{
\dontrun{
calcOutput("LivestockProductivity")
}
}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}, \code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Isabelle Weindl, Lavinia Baumstark, Stephen Wirth
}
