% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLivstFactorCostCalib.R
\name{calcLivstFactorCostCalib}
\alias{calcLivstFactorCostCalib}
\title{calcLivstFactorCostCalib}
\usage{
calcLivstFactorCostCalib()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calibrates the global regression for livestock factor requirements to iso level by adjusting the
intercept
}
\examples{
\dontrun{
calcOutput("LivstFactorCostCalib")
}
}
\author{
Debbora Leip
}
