% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmL.R
\name{calcLPJmL}
\alias{calcLPJmL}
\title{calcLPJmL}
\usage{
calcLPJmL(
  version = "LPJmL4",
  climatetype = "CRU_4",
  subtype = "soilc",
  subdata = NULL,
  time = "raw",
  averaging_range = NULL,
  dof = NULL,
  harmonize_baseline = FALSE,
  ref_year = "y2015",
  limited = TRUE,
  hard_cut = FALSE,
  selectyears = "all"
)
}
\arguments{
\item{version}{Switch between LPJmL4 and LPJmL4}

\item{climatetype}{Switch between different climate scenarios (default: "CRU_4")}

\item{subtype}{Switch between different lpjml input as specified in readLPJmL}

\item{subdata}{Switch between data dimension subitems}

\item{time}{average, spline or raw (default)}

\item{averaging_range}{just specify for time=="average": number of time steps to average}

\item{dof}{just specify for time=="spline": degrees of freedom}

\item{harmonize_baseline}{FALSE (default) nothing happens, if a baseline is specified here data is harmonized to that baseline (from ref_year on)}

\item{ref_year}{just specify for harmonize_baseline != FALSE : Reference year}

\item{limited}{just specify for harmonize_baseline != FALSE : if TRUE limited approached is used}

\item{hard_cut}{just specify for harmonize_baseline != FALSE : use hard cut instead of multiplicative factor}

\item{selectyears}{defaults to all years available}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Handle LPJmL data and its time behaviour (averaging, approximation, harmonizing to baseline)
}
\examples{

\dontrun{ 
calcOutput("LPJmL", version="LPJmL4", climatetype="CRU_4", subtype="soilc", aggregate=FALSE)
}
}
\seealso{
\code{\link{readLPJmL}}
}
\author{
Kristine Karstens, Felicitas Beier
}
