% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmL_new.R
\name{calcLPJmL_new}
\alias{calcLPJmL_new}
\title{calcLPJmL_new}
\usage{
calcLPJmL_new(
  version = "LPJmL4_for_MAgPIE_44ac93de",
  climatetype = "MRI-ESM2-0:ssp370",
  subtype = "soilc",
  subdata = NULL,
  stage = "harmonized2020"
)
}
\arguments{
\item{version}{Switch between LPJmL versions (including addons for further version specification)}

\item{climatetype}{Switch between different climate scenarios}

\item{subtype}{Switch between different lpjml input as specified in readLPJmL}

\item{subdata}{Switch between data dimension subitems}

\item{stage}{Degree of processing: raw, smoothed, harmonized, harmonized2020}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Handle LPJmL data and its time behaviour (smoothing and harmonizing to baseline)
}
\examples{
\dontrun{
calcOutput("LPJmL_new", subtype = "soilc", aggregate = FALSE)
}

}
\seealso{
\code{\link{readLPJmL}}
}
\author{
Kristine Karstens, Felicitas Beier
}
