% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLPJmLClimateInput.R
\name{calcLPJmLClimateInput}
\alias{calcLPJmLClimateInput}
\title{calcLPJmLClimateInput}
\usage{
calcLPJmLClimateInput(
  climatetype = "MRI-ESM2-0:ssp370",
  variable = "temperature:annualMean",
  stage = "harmonized2020",
  lpjmlVersion = "LPJmL4_for_MAgPIE_44ac93de"
)
}
\arguments{
\item{climatetype}{Switch between different climate scenario}

\item{variable}{Switch between different climate inputs and temporal resolution}

\item{stage}{Degree of processing: raw, smoothed - raw or smoothed data from 1930|1951
raw1901, smoothed1901 - raw or smoothed data from 1901
harmonized, harmonized2020 - based on toolLPJmLVersion}

\item{lpjmlVersion}{LPJmL Version hand over}
}
\value{
magpie object in cellular resolution
}
\description{
Handle LPJmL climate input data and its time behaviour
(smoothing and harmonizing to baseline)
}
\examples{
\dontrun{
calcOutput("LPJmLClimateInput",
           climatetype = "MRI-ESM2-0:ssp370",
           variable = "temperature:annualMean")
}

}
\author{
Marcos Alves, Kristine Karstens, Felicitas Beier
}
