% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH2MAgPIE.R
\name{calcLUH2MAgPIE}
\alias{calcLUH2MAgPIE}
\title{calcLUH2MAgPIE}
\usage{
calcLUH2MAgPIE(
  share = "total",
  bioenergy = "ignore",
  rice = "non_flooded",
  selectyears = "past",
  missing = "ignore"
)
}
\arguments{
\item{share}{total (for total numbers), LUHofMAG (for share of LUH within kcr types),
MAGofLUH (for share of kcr within LUH types)}

\item{bioenergy}{"ignore": 0 for share and totals,
"fix": fixes betr and begr shares in LUHofMAG to 1 for c3per and c4per}

\item{rice}{rice category: "non_flooded" or "total"}

\item{selectyears}{years to be returned (default: "past")}

\item{missing}{"ignore" will leave data as is,
"fill" will add proxy values for data gaps of FAO}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description
}
\description{
Calculates the real aggregation of LUH croptypes to MAgPIE croptypes
out of LUH2FAO and FAO2MAgPIE mappings
}
\examples{
\dontrun{
calcOutput("LUH2MAgPIE")
}

}
\author{
Kristine Karstens, Felicitas Beier
}
