% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcLUH2v2.R
\name{calcLUH2v2}
\alias{calcLUH2v2}
\title{calcLUH2v2}
\usage{
calcLUH2v2(
  landuse_types = "magpie",
  irrigation = FALSE,
  cellular = FALSE,
  cells = "magpiecell",
  selectyears = "past"
)
}
\arguments{
\item{landuse_types}{magpie: magpie landuse classes,
LUH2v2: original landuse classes
flooded: flooded areas as reported by LUH}

\item{irrigation}{if true: areas are returned separated by irrigated and rainfed,
if false: total areas}

\item{cellular}{if true: dataset is returned on 0.5 degree resolution}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420)}

\item{selectyears}{years to be returned (default: "past")}
}
\value{
List of magpie objects with results on country level,
        weight on country level, unit and description
}
\description{
Integrates the LUH2v2 landuse-dataset
}
\examples{
\dontrun{
calcOutput("LUH2v2")
}
}
\seealso{
\code{\link{calcLanduseInitialisation}}
}
\author{
Benjamin Leon Bodirsky, Florian Humpenoeder, Jens Heinke, Felicitas Beier
}
