% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMACCsN2O.R
\name{calcMACCsN2O}
\alias{calcMACCsN2O}
\title{Calculation of N2O MAC curves of Energy Industry and Landuse}
\usage{
calcMACCsN2O(sector = "all", source = "ImageMacc")
}
\arguments{
\item{sector}{"all" or "landuse"; "all"" includes energy_industry and landuse}

\item{source}{"ImageMacc" or "PBL_MACC_2019"}
}
\value{
MAgPIE object
}
\description{
Calculation of the N2O relative costcurves (subtypes: Energy Industry and
Landuse) weighted by the baseline emissions. Sources: N2O Transport, N2O
Adipic acid production, N2O Nitric acid production, N2O Fertilizer, N2O
Animal waste, N2O Domestic sewage divided in classes 1-201.
}
\examples{
\dontrun{
calcOutput("MACCsN2O")
}
}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}, \code{\link[=readImageMacc]{readImageMacc()}},
\code{\link[=convertImageMacc]{convertImageMacc()}}
}
\author{
Nele Steinmetz, Florian Humpenoeder, Michael Windisch
}
