% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcManureRecyclingCroplandPast.R
\name{calcManureRecyclingCroplandPast}
\alias{calcManureRecyclingCroplandPast}
\title{calcManureRecyclingCroplandPast}
\usage{
calcManureRecyclingCroplandPast(
  products = "sum",
  cellular = FALSE,
  cells = "lpjcell"
)
}
\arguments{
\item{products}{"sum" (default) or "kli"}

\item{cellular}{if TRUE value is calculate and returned (set aggregate to FALSE!) on cellular level}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420)}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates manure recycling to cropland based on excretions, animal waste management types
(and their shares per country) and emission factors for nitrogenous emissions in livestock confinements
}
\examples{
\dontrun{
calcOutput("ManureRecyclingCroplandPast")
}
}
\seealso{
\code{\link[=calcExcretion]{calcExcretion()}}
}
\author{
Benjamin Leon Bodirsky, Kristine Karstens
}
