% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticropping.R
\name{calcMulticropping}
\alias{calcMulticropping}
\title{calcMulticropping}
\usage{
calcMulticropping(extend_future = FALSE)
}
\arguments{
\item{extend_future}{if TRUE}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the ratio between area harvested and physical cropland area. Can be larger or smaller, depending on fallow land and double cropping.
}
\examples{

\dontrun{
calcOutput("Multicropping")
}

}
\seealso{
\code{\link[=calcFAOLand]{calcFAOLand()}},
\code{\link[=calcCroparea]{calcCroparea()}}
}
\author{
Benjamin Leon Bodirsky, David Chen
}
