% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMulticroppingSuitability.R
\name{calcMulticroppingSuitability}
\alias{calcMulticroppingSuitability}
\title{calcMulticroppingSuitability}
\usage{
calcMulticroppingSuitability(
  selectyears,
  lpjml,
  climatetype,
  suitability = "endogenous",
  sectoral = "kcr"
)
}
\arguments{
\item{selectyears}{Years to be returned}

\item{lpjml}{LPJmL version required for respective inputs: natveg or crop}

\item{climatetype}{Switch between different climate scenarios or
historical baseline "GSWP3-W5E5:historical"}

\item{suitability}{"endogenous": suitability for multiple cropping determined
by rules based on grass and crop productivity
"exogenous": suitability for multiple cropping given by
GAEZ data set}

\item{sectoral}{"kcr" MAgPIE crops, and "lpj" LPJmL crops}
}
\value{
magpie object in cellular resolution
}
\description{
Calculates which grid cells are potentially suitable for
multiple cropping activities under rainfed and irrigated conditions.
Calculation is based on the length of the growing period determined by
monthly grassland gross primary production (GPP).
}
\examples{
\dontrun{
calcOutput("MulticroppingSuitability", aggregate = FALSE)
}

}
\author{
Felicitas Beier, Jens Heinke
}
