% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcMultipleCroppingZones.R
\name{calcMultipleCroppingZones}
\alias{calcMultipleCroppingZones}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{  calcMultipleCroppingZones
}\if{html}{\out{</div>}}}
\usage{
calcMultipleCroppingZones(layers = 2)
}
\arguments{
\item{layers}{8 for original GAEZ layers,
3 for aggregated multiple cropping zones with
1 = single cropping, 2 = double cropping, 3 = triple cropping
2 for aggregated boolean multicropping potential with
0 = no multicropping (single cropping), 1 = multiple cropping}
}
\value{
magpie object in cellular resolution
}
\description{
This function returns multiple cropping zones at 0.5 degree resolution
}
\examples{
\dontrun{
calcOutput("MultipleCroppingZones", layers = 3, aggregate = FALSE)
}

}
\author{
Felicitas Beier
}
