% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenBudgetNonagland.R
\name{calcNitrogenBudgetNonagland}
\alias{calcNitrogenBudgetNonagland}
\title{calcNitrogenBudgetNonagland}
\usage{
calcNitrogenBudgetNonagland(
  deposition = "CEDS",
  max_nue = 0.95,
  cellular = FALSE
)
}
\arguments{
\item{deposition}{if FALSE, deposition is not accounted for in the distribution.
Use FALSE to avoid circularities in calcNitrogenBudget}

\item{max_nue}{NULL or a numeric value. if numeric, an additional N balanceflow is included that takes
care that the nitrogen use efficiency does not exceed the numeric value in balanceflow.}

\item{cellular}{TRUE returns output on 0.5 degree grid}
}
\value{
Nitrogen Budgets for Non-agricultural land
}
\description{
Calculates Nitrogen Budgets for Non-agricultural land.
}
\examples{
\dontrun{
calcOutput("NitrogenBudgetNonagland")
}
}
\author{
Benjamin Leon Bodirsky
}
