% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationBNF.R
\name{calcNitrogenFixationBNF}
\alias{calcNitrogenFixationBNF}
\title{calcNitrogenFixationPast}
\usage{
calcNitrogenFixationBNF(
  sum_plantparts = TRUE,
  cellular = FALSE,
  irrigation = FALSE
)
}
\arguments{
\item{sum_plantparts}{if false, crop residues, belowground residues and harvested organ are reported separately}

\item{cellular}{cellular estimates optional}

\item{irrigation}{if TRUE, distinguishes irrigated and non-irrigated crops}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates fixation from nitrogen-fixing crops
}
\examples{
\dontrun{
calcOutput("calcNitrogenFixationBNF")
}
}
\seealso{
\code{\link[=calcNitrogenFixationPast]{calcNitrogenFixationPast()}}
}
\author{
Benjamin Leon Bodirsky
}
