% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationFreeliving.R
\name{calcNitrogenFixationFreeliving}
\alias{calcNitrogenFixationFreeliving}
\title{calcNitrogenFixationFreeliving}
\usage{
calcNitrogenFixationFreeliving(cellular = FALSE, irrigation = FALSE)
}
\arguments{
\item{cellular}{cellular estimates optional}

\item{irrigation}{if TRUE, distinguishes irrigated and non-irrigated crops}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates fixation from freeliving bacteria
}
\examples{
\dontrun{
calcOutput("NitrogenFixationFreeliving")
}
}
\author{
Benjamin Leon Bodirsky
}
