% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationRateNatural.R
\name{calcNitrogenFixationRateNatural}
\alias{calcNitrogenFixationRateNatural}
\title{calcNitrogenFixationRateNatural}
\usage{
calcNitrogenFixationRateNatural(cells = "lpjcell")
}
\arguments{
\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
List of magpie objects with results on global level, empty weight, unit and description.
}
\description{
calculates fixation rates from natural ecosystems based on evapostranspiration
}
\examples{
\dontrun{
calcOutput("NitrogenFixationRateNatural")
}

}
\seealso{
\code{\link[=calcNitrogenFixationPast]{calcNitrogenFixationPast()}}
\code{\link[=readHerridge]{readHerridge()}}
}
\author{
Benjamin Leon Bodirsky
}
