% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationSymbiotic.R
\name{calcNitrogenFixationSymbiotic}
\alias{calcNitrogenFixationSymbiotic}
\title{calcNitrogenFixationSymbiotic}
\usage{
calcNitrogenFixationSymbiotic(cellular = FALSE, irrigation = FALSE)
}
\arguments{
\item{cellular}{cellular estimates optional}

\item{irrigation}{if TRUE, distinguishes irrigated and non-irrigated crops}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates fixation from nitrogen-fixing crops
}
\examples{
\dontrun{
calcOutput("NitrogenFixationSymbiotic")
}
}
\author{
Benjamin Leon Bodirsky
}
