% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNonMAgPIEFactorCosts.R
\name{calcNonMAgPIEFactorCosts}
\alias{calcNonMAgPIEFactorCosts}
\title{calcNonMAgPIEFactorCosts}
\usage{
calcNonMAgPIEFactorCosts(
  subtype = "subsidies",
  aggSubsidies = FALSE,
  extrapolate = TRUE
)
}
\arguments{
\item{subtype}{either factor cost share of "subsidies" (which don't enter MAgPIE labor costs as they should not
affect prices), or of "missingVoP" (which refers to livestock VoP that can't be mapped to MAgPIE livestock
categories, i.e. wool, beeswax, honey, silk-worms)}

\item{aggSubsidies}{boolean: if subtype is "subsidies", should crop and livestock subsidies be reported separately
or as aggregate?}

\item{extrapolate}{boolean: should values be extrapolate (by keeping constant) until 2150?}
}
\value{
magpie object. in mio. USDMER05
}
\description{
Calculates factor costs that should affect agricultural employment but are not included in MAgPIE
factor costs
}
\examples{
\dontrun{
a <- calcOutput("NonMAgPIEFactorCosts", subtype = "subsidies")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Debbora Leip
}
