% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNuePasture.R
\name{calcNuePasture}
\alias{calcNuePasture}
\title{calcNuePasture}
\usage{
calcNuePasture(cellular = FALSE, maccbase = TRUE)
}
\arguments{
\item{cellular}{cellular disaggreagation or national values}

\item{maccbase}{whether future scenarios should be expressed as base efficiency, excluding additional macc improvemetns (new default)}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates the soil nitrogen uptake efficiency of pastures. This is the nitrogen taken up from the soil (N in crop biomass minus biological fixation minus seed N) divided by the soil N inputs (fertilizer, manure etc). For the future, NUE scenarios are added.
}
\examples{

\dontrun{
calcOutput("NuePasture")
}

}
\seealso{
\code{\link[=calcSNUpE]{calcSNUpE()}}
\code{\link[=calcNitrogenBudgetPasture]{calcNitrogenBudgetPasture()}}
}
\author{
Benjamin Leon Bodirsky
}
