% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlantationCellular.R
\name{calcPlantationCellular}
\alias{calcPlantationCellular}
\title{calcPlantationCellular}
\usage{
calcPlantationCellular()
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates the cellular plantation area based on carbon densities
}
\examples{

\dontrun{
calcOutput("PlantationCellular")
}
}
\author{
Abhijeet Mishra
}
