% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPlantEstablishCalib.R
\name{calcPlantEstablishCalib}
\alias{calcPlantEstablishCalib}
\title{calcPlantEstablishCalib}
\usage{
calcPlantEstablishCalib()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the calibration factors for plantation establishment globally
}
\examples{

\dontrun{ 
calcOutput("PlantEstablishCalib",aggregate=TRUE)
}
}
\author{
Abhijeet Mishra
}
