% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulation.R
\name{calcPopulation}
\alias{calcPopulation}
\title{calcPopulation}
\usage{
calcPopulation(
  PopulationCalib = c("past_grPEAP_grFuture", "Ariadne"),
  PopulationPast = c("WDI_completed", "Eurostat_WDI_completed"),
  PopulationFuture = c("SSP2018Update_completed_bezierOut",
    "SSP2Ariadne_completed_bezierOut"),
  FiveYearSteps = TRUE,
  naming = "indicator_scenario"
)
}
\arguments{
\item{PopulationCalib}{to what should be calibrated? past, future or a transition?}

\item{PopulationPast}{population past data source}

\item{PopulationFuture}{population future data source}

\item{FiveYearSteps}{Only five year steps if TRUE, FALSE returns years from source data}

\item{naming}{naming scheme}
}
\value{
Population in millions.
}
\description{
Merges time series of population for the past and present. Different sources
are available and can be selected in the madrat config. See
\code{\link{calcPopulationPast}} for past datasets, and
\code{\link{calcPopulationFuture}} for future datasets. The time series are
merged via the growth rates. The first year of the future scenarios
determines the merging point. All data is calibrated either to the "past" or
the "future" dataset as specified in "getConfig()$calc$PopulationCalib".
}
\seealso{
\code{\link{calcPopulationPast}},\code{\link{calcPopulationFuture}}
}
\author{
Lavinia Baumstark, Benjamin Bodirsky, Johannes Koch
}
