% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPopulationFuture.R
\name{calcPopulationFuture}
\alias{calcPopulationFuture}
\title{calcPopulationFuture}
\usage{
calcPopulationFuture(PopulationFuture = "SSP2018Update_completed_bezierOut")
}
\arguments{
\item{PopulationFuture}{population future data source}
}
\value{
Population in millions.
}
\description{
Calculates a time series of Population. Different sources are available and
can be selected in the madrat config (getConfig()$calc$PopulationPast):
\itemize{ \item \code{"IIASApop"}: Source: IIASA? Lavinia?  \item
\code{"IIASApop"}: Source: Lavinia? }
}
\author{
Lavinia Baumstark, Benjamin Bodirsky, Johannes Koch
}
