% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcPricesProducer.R
\name{calcPricesProducer}
\alias{calcPricesProducer}
\title{calcPricesProducer}
\usage{
calcPricesProducer(products = "kcr", calculation = "VoP")
}
\arguments{
\item{products}{either "kcr" or "kcl"}

\item{calculation}{type of calculation "FAO" (directly reads the data), "VoP"
calculates as VoP/Production, only "FAO" available for "kli" products.}
}
\value{
magpie object. prices in year specific annual
}
\description{
producer prices for agricultural products. 05USDppp/tDM
}
\examples{
\dontrun{
a <- calcOutput("PricesProducer")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Edna J. Molina Bacca
}
