% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProtectArea.R
\name{calcProtectArea}
\alias{calcProtectArea}
\title{calcProtectArea}
\usage{
calcProtectArea(cells = "magpiecell", bhifl = TRUE)
}
\arguments{
\item{cells}{number of cells of landmask (select "magpiecell" for 59199 cells or "lpjcell" for 67420 cells)}

\item{bhifl}{should be TRUE (including BH_IFL scenario)
for cellular preprocessing revisions > 4.65}
}
\value{
magpie object in cellular resolution with different protection scenarios
}
\description{
Function extracts conservation protected area
}
\examples{
\dontrun{
calcOutput("ProtectArea", aggregate = FALSE)
}

}
\author{
Felicitas Beier, David Chen
}
