% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRegFactorShare.R
\name{calcRegFactorShare}
\alias{calcRegFactorShare}
\title{calcRegFactorShare}
\usage{
calcRegFactorShare(datasource = "USDA", factor = "cap")
}
\arguments{
\item{datasource}{Only USDA available}

\item{factor}{"lab" for Labour and "cap" for capital}
}
\value{
MAgPIE object at global level with slope and intersect as items
}
\description{
This function calculates the regression parameters (a and b) for the function Share=a*log10(GDP)+b. Where share is the adjusted share between capital and labour.
}
\examples{
\dontrun{
calcOutput("calcRegFactorShare")
}


}
\seealso{
\code{\link[=calcOutput]{calcOutput()}},\code{\link[=calcFactorIntensity]{calcFactorIntensity()}}
}
\author{
Debbora Leip, Edna J. Molina Bacca
}
