% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcResFieldBalancePast.R
\name{calcResFieldBalancePast}
\alias{calcResFieldBalancePast}
\title{calcResFieldBalancePast}
\usage{
calcResFieldBalancePast(
  cellular = FALSE,
  products = "sum",
  scenario = "default"
)
}
\arguments{
\item{cellular}{If TRUE calculation and output on cellular level}

\item{products}{"sum" (default) or "kres"}

\item{scenario}{define scenario switch for sensititvy analysis
for historical SOC budget}
}
\value{
data
}
\description{
Calculates data for aboveground and belowground residues production with other usage
}
\examples{
\dontrun{
calcOutput("ResFieldBalancePast")
}

}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}, \code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Benjamin Bodirsky
}
