% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRicearea.R
\name{calcRicearea}
\alias{calcRicearea}
\title{calcRicearea}
\usage{
calcRicearea(cellular = FALSE, cells = "magpiecell", share = TRUE)
}
\arguments{
\item{cellular}{If TRUE: calculates cellular rice area}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420)}

\item{share}{If TRUE: non-flooded share is returned.
If FALSE: rice area (flooded, non-flooded, total) in Mha is returned}
}
\value{
rice areas or rice area shares of flooded and non-flooded category
}
\description{
calculates rice area based on LUH flooded areas and
physical rice areas reported by FAOSTAT.
}
\author{
Felicitas Beier, Kristine Karstens
}
