% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcSOMlossN.R
\name{calcSOMlossN}
\alias{calcSOMlossN}
\title{calcSOMlossN}
\usage{
calcSOMlossN(cellular = FALSE, cells = "lpjcell")
}
\arguments{
\item{cellular}{if TRUE cellular level is returned}

\item{cells}{"magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
List of magpie object with results on country or cellular level,
weight on cellular level, unit and description.
}
\description{
calculates effect on N  from Soil Organic Matter loss
}
\examples{
\dontrun{
calcOutput("SOMlossN")
}

}
\author{
Benjamin Leon Bodirsky,
}
