% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcStorage.R
\name{calcStorage}
\alias{calcStorage}
\title{calcStorage}
\usage{
calcStorage(datasource = "FAO")
}
\arguments{
\item{datasource}{Options of the source of data, currently only FAO:  \code{FAO}.}
}
\value{
List with a magpie object with stock level in tonnes
}
\description{
calculates stock levels of magpie commodities based on FAO data.
FAO data only has stock variation so we assume lowest level for each product/country as 0 level
USDA data to be potentially included in future second data source
}
\examples{

\dontrun{
calcOutput("Storage", datasource="FAO")
}


}
\author{
David Chen
}
