% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTemperature.R
\name{calcTemperature}
\alias{calcTemperature}
\title{calcTemperature}
\usage{
calcTemperature(landusetypes = "all", months = FALSE, convert = TRUE)
}
\arguments{
\item{landusetypes}{all or only one (to save computation memory)}

\item{months}{FALSE for yearly average, TRUE for monthly values}

\item{convert}{FALSE for raw values of temperature,
TRUE add temperature of 15 degrees for countries
without observations or land mass.}
}
\value{
List of magpie object with results on country level, weight on country level, unit and description.
}
\description{
calculates average monthly temperature on different landuse types
}
\examples{
\dontrun{
calcOutput("Temperature")
}
}
\author{
Benjamin Leon Bodirsky
}
