% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcTimberTradeBilateral.R
\name{calcTimberTradeBilateral}
\alias{calcTimberTradeBilateral}
\title{calcTimberTradeBilateral}
\usage{
calcTimberTradeBilateral()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
A  very rough disaggregation of timber demand to bilateral trade flows
}
\examples{
\dontrun{
calcOutput("TimberTradeBilateral")
}
}
\seealso{
\code{\link[mrfaocore:calcFAOmassbalance_pre]{mrfaocore::calcFAOmassbalance_pre()}}
}
\author{
David M Chen
}
