% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVoP_AFF.R
\name{calcVoP_AFF}
\alias{calcVoP_AFF}
\title{calcVoP_AFF}
\usage{
calcVoP_AFF()
}
\value{
magpie object. in mio. 05USD ppp units
}
\description{
Calculates the overall value of production of the agriculture,
forestry and fisheries sectors. Forestry and Fisheries are calculated from exports values.
}
\examples{
\dontrun{
a <- calcOutput("VoP_AFF")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Edna J. Molina Bacca
}
