% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVoP_crops.R
\name{calcVoP_crops}
\alias{calcVoP_crops}
\title{calcVoP_crops}
\usage{
calcVoP_crops(output = "absolute")
}
\arguments{
\item{output}{defines if output should be given as an "absolute" value or
as a "fraction" of the overall value of production.}
}
\value{
magpie object. in mio. USD05 MER or fraction
}
\description{
Calculates the value of production of individual production items or
its fraction compared to overall Value of Production (Agriculture,Fish,Forestry).
}
\examples{
\dontrun{
a <- calcOutput("VoP_crops")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Edna J. Molina Bacca
}
