% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVoP_livst.R
\name{calcVoP_livst}
\alias{calcVoP_livst}
\title{calcVoP_livst}
\usage{
calcVoP_livst(other = FALSE)
}
\arguments{
\item{other}{boolean: should FAO livestock categories that can't be matched to MAgPIE categories (i.e. beeswax, wool,
silkworms, and honey) be reported as "livst_other"?}
}
\value{
magpie object. in mio. USDMER05
}
\description{
Calculates the value of production of individual livestock categories
}
\examples{
\dontrun{
a <- calcOutput("VoP_livst")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Debbora Leip
}
