% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcVoPAFF.R
\name{calcVoPAFF}
\alias{calcVoPAFF}
\title{calcVoPAFF}
\usage{
calcVoPAFF()
}
\value{
magpie object. in mio. 05USD MER units
}
\description{
Calculates the overall value of production of the agriculture,
forestry and fisheries sectors. Forestry and Fisheries are calculated from exports values.
}
\examples{
\dontrun{
a <- calcOutput("VoPAFF")
}

}
\seealso{
\code{\link[=calcOutput]{calcOutput()}}
}
\author{
Edna J. Molina Bacca, Debbora Leip
}
