% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWBGEM.R
\name{calcWBGEM}
\alias{calcWBGEM}
\title{calcWBGEM}
\usage{
calcWBGEM()
}
\value{
magpie object of time series of world price of commodities
}
\description{
recover world price of commodity in terms of real 2005 USD per metric ton
}
\examples{
\dontrun{
calcOutput("WBGEM")
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}},
\code{\link[=readWBGEM]{readWBGEM()}}
}
\author{
Xiaoxi Wang
}
