% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWeeklyHoursILO.R
\name{calcWeeklyHoursILO}
\alias{calcWeeklyHoursILO}
\title{calcWeeklyHoursILO}
\usage{
calcWeeklyHoursILO(projections = FALSE)
}
\arguments{
\item{projections}{boolean, should weekly hours be projected up to 2100}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates complete dataset of mean weekly hours worked by people employed in agriculture, forestry and
fishery based on ILO dataset
}
\examples{
\dontrun{
calcOutput("WeeklyHoursILO")
}
}
\author{
Debbora Leip
}
