% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcWoodDensity.R
\name{calcWoodDensity}
\alias{calcWoodDensity}
\title{calcWoodDensity}
\usage{
calcWoodDensity()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates country-level basic wood density (tDM per m3) based on
IPCC 2006 Vol 4 Chapter 12 Table 12.4 values and Koeppen-Geiger climate classification.
Density values: Tropical (A) = 0.59, Arid (B) = 0.52, Temperate (C) = 0.45,
Continental (D) = 0.45, Polar (E) = 0.45.
}
\examples{
\dontrun{
calcOutput("WoodDensity")
}
}
\author{
Florian Humpenoeder
}
