% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertBodirsky2018.R
\name{convertBodirsky2018}
\alias{convertBodirsky2018}
\title{convertBodirsky2018}
\usage{
convertBodirsky2018(x, subtype)
}
\arguments{
\item{x}{magpie object from read function}

\item{subtype}{bmi_share, demand_regression, intake_regression or bodyheight_regression for the estimated regression paramters of different regressions. scenarios for scenario projections.}
}
\value{
magpie object
}
\description{
Reads in regression parameters estimated using mrregression, and reads in some scenario
based on Bodirsky et al not yet published
}
\seealso{
\code{\link[=readBodirsky2018]{readBodirsky2018()}}
}
