% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCEDS2021.R
\name{convertCEDS2021}
\alias{convertCEDS2021}
\title{convertCEDS2021}
\usage{
convertCEDS2021(x)
}
\arguments{
\item{x}{magpie object from source function}
}
\value{
MAgPIE object
}
\description{
converts emission data from the CEDS database
}
\author{
Benjamin Leon Bodirsky, David Klein
}
