% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCEDS2024.R
\name{convertCEDS2024}
\alias{convertCEDS2024}
\title{convertCEDS2024}
\usage{
convertCEDS2024(x)
}
\arguments{
\item{x}{magpie object from source function}
}
\value{
MAgPIE object
}
\description{
converts emission data from the CEDS database
}
\author{
Benjamin Leon Bodirsky, David Klein
}
