% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCEDS2025.R
\name{convertCEDS2025}
\alias{convertCEDS2025}
\title{convertCEDS2025}
\usage{
convertCEDS2025(x)
}
\arguments{
\item{x}{magpie object from source function}
}
\value{
MAgPIE object
}
\description{
converts emission data from the CEDS database
}
\author{
Benjamin Leon Bodirsky, David Klein
}
