% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCEDS_CMIP7.R
\name{convertCEDS_CMIP7}
\alias{convertCEDS_CMIP7}
\title{convertCEDS_CMIP7}
\usage{
convertCEDS_CMIP7(x)
}
\arguments{
\item{x}{magpie object from source function}
}
\value{
MAgPIE object
}
\description{
converts harmonized emission data used for CMIP7, global data
is removed here, use \code{readSource(CEDS_CMIP7, convert = FALSE)} to obtain it.
}
\author{
Pascal Weigmann
}
