% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertCMIP7_CEDS.R
\name{convertCMIP7_CEDS}
\alias{convertCMIP7_CEDS}
\title{convertCMIP7_CEDS}
\usage{
convertCMIP7_CEDS(x)
}
\arguments{
\item{x}{magpie object from source function}
}
\value{
MAgPIE object
}
\description{
converts harmonized emission data used for CMIP7, global data
is removed here, use \code{readSource(CMIP7_CEDS, convert = FALSE)} to obtain it.
}
\author{
Pascal Weigmann
}
