% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertDomingues.R
\name{convertDomingues}
\alias{convertDomingues}
\title{convertDomingues}
\usage{
convertDomingues(x)
}
\arguments{
\item{x}{unconverted magpie object from read-script}
}
\value{
Data as MAgPIE object with common country list
}
\description{
fills missing countries in Domingues data with NA, and transforms
cost units from 2018 US$/GJ to 2017 US$/GJ
}
\author{
Eva Bleidorn
}
