% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertEurostat.R
\name{convertEurostat}
\alias{convertEurostat}
\title{Converts Eurostat historical emissions}
\usage{
convertEurostat(x, subtype)
}
\arguments{
\item{x}{MAgPIE object to be converted}

\item{subtype}{emissions for original eurostat emissions split, MACCemi for MACC historical emissions, or
sectorEmi for sector specific emissions}
}
\value{
A MAgPIE object containing the Eurostat historical emissions (MtCO2)
}
\description{
Converts Eurostat historical emissions
}
\examples{
\dontrun{
a <- convertEurostat(x, subtype = "emissions")
}

}
\author{
Renato Rodrigues
}
