% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFAO_TradeMatrix.R
\name{convertFAO_TradeMatrix}
\alias{convertFAO_TradeMatrix}
\title{Convert FAO_TradeMatrix}
\usage{
convertFAO_TradeMatrix(x, subtype)
}
\arguments{
\item{x}{output from read function}

\item{subtype}{subsets of the detailed trade matrix to read in. Very large csv needs to be read in chunks
separated by export/import quantities and values, as well as kcr, kli and kothers (not in kcr nor kli)
Options are all combinations of c("import_value", "import_qty", "export_value", "export_quantity") X c("kcr", "kli", "kothers"))
import is import side reporting while export is export-sde reporting}
}
\value{
FAO data as MAgPIE object
}
\description{
Convert FAOSTAT detailed trade matrix.
FAOSTAT does not balance or harmonize the import/export side reporting, effectively creating 2 slightly different datasets.
Furthermore, in terms of trade value, exporters are "usuallY" reporting FOB, while importers report CIF. Difference in value,
given identical qty, is thus the transport margin.
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
David C
}
