% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFeedModel.R
\name{convertFeedModel}
\alias{convertFeedModel}
\title{Convert FeedModel data}
\usage{
convertFeedModel(x, subtype = "FeedBaskets")
}
\arguments{
\item{x}{MAgPIE object containing production system distribution and feed basket data.}

\item{subtype}{Available subtypes: "ProdSysRatio", "FeedBaskets" and "FeedBasketsDetailed"}
}
\value{
Production system distribution and feed basket data as MAgPIE object on ISO country level
}
\description{
Convert production system distribution and feed basket data to ISO country level.
}
\author{
Isabelle Weindl, Jan Philipp Dietrich
}
