% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertFRA2020.R
\name{convertFRA2020}
\alias{convertFRA2020}
\title{Convert FRA 2020 data}
\usage{
convertFRA2020(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{The FAO FRA 2020 subtype.}
}
\value{
Data as MAgPIE object with common country list
}
\description{
Convert FRA 2020 data
}
\examples{
\dontrun{
a <- readSource("FRA2020", "growing_stock", convert = TRUE)
}
}
\seealso{
\code{\link[=readSource]{readSource()}},
}
\author{
Abhijeet Mishra
}
