% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGGCMIHarvestAreaMask.r
\name{convertGGCMIHarvestAreaMask}
\alias{convertGGCMIHarvestAreaMask}
\title{convertGGCMIHarvestAreaMask}
\usage{
convertGGCMIHarvestAreaMask(x)
}
\arguments{
\item{x}{object coming from read function}
}
\value{
MAgPIE object with the requested data
}
\description{
Reads in GGCMI fraction of Harvested Area masks for rice 1 and rice 2
(other crops available too, see path in download fucntion)
}
\author{
David M Chen
}
