% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGTAP81.R
\name{convertGTAP81}
\alias{convertGTAP81}
\title{convertGTAP81}
\usage{
convertGTAP81(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{GTAP header that should be read}
}
\value{
Data as MAgPIE object with common country list
}
\description{
disaggregates country groups from GTAP according to GDP share
}
\examples{
\dontrun{
a <- readSource("GTAP81", "SF01", convert = TRUE)
}

}
\author{
Debbora Leip
}
