% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertGTAPv8v9.R
\name{convertGTAPv8v9}
\alias{convertGTAPv8v9}
\title{convertGTAPv8v9}
\usage{
convertGTAPv8v9(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{GTAP header that should be read}
}
\value{
Data as MAgPIE object with common country list
}
\description{
disaggregates country groups from GTAP according to GDP share
}
\examples{
\dontrun{
a <- readSource("convertGTAPv8v9", "81:SF01", convert = TRUE)
}

}
\author{
Debbora Leip, David M. Chen
}
