% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertICP2017.R
\name{convertICP2017}
\alias{convertICP2017}
\title{convertICP2017}
\usage{
convertICP2017(x, subtype = "per_cap_expMER")
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{data subtype needed. Either "priceLevel", or "per_cap_exp"}
}
\value{
magpie object of relative price levels (world = 100) or per capita expenditure (USD17 MER)
}
\description{
converts data of World Bank ICP round, downloaded from
here: https://databank.worldbank.org/source/icp-2017
mainly a currency conversion to MER05. Assume that in the original dataset,
2011 values are in 2011MER and 2017 in 2017MER, given the ICP rounds
}
\examples{
\dontrun{
a <- convertSource("ICP2017", "per_cap_exp")
}

}
\author{
David M Chen
}
