% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_EEI.R
\name{convertIEA_EEI}
\alias{convertIEA_EEI}
\title{Convert IEA End Uses and Efficiency Indicators data to data on ISO country level.}
\usage{
convertIEA_EEI(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\description{
Convert IEA End Uses and Efficiency Indicators data to data on ISO country level.
}
\author{
Falk Benke
}
