% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIEA_EnergyEfficiencyIndicators.R
\name{convertIEA_EnergyEfficiencyIndicators}
\alias{convertIEA_EnergyEfficiencyIndicators}
\title{Convert IEA End Uses and Efficiency Indicators data to data on ISO country level.}
\usage{
convertIEA_EnergyEfficiencyIndicators(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\description{
Convert IEA End Uses and Efficiency Indicators data to data on ISO country level.
}
\author{
Falk Benke
}
