% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIFPRIsubsidy.R
\name{convertIFPRIsubsidy}
\alias{convertIFPRIsubsidy}
\title{convertIFPRIsubsidy}
\usage{
convertIFPRIsubsidy(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object of agricultural subsidies
}
\description{
disaggregates EU subsidies for crops and livestock from IFPRI table to country level
}
\examples{
\dontrun{
a <- readSource("IFPRIsubsidy", convert = TRUE)
}

}
\author{
Debbora Leip
}
