% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertILOSTAT.R
\name{convertILOSTAT}
\alias{convertILOSTAT}
\title{convertILOSTAT}
\usage{
convertILOSTAT(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{Type of ILOSTAT data that should be read}
}
\value{
Data as MAgPIE object with common country list
}
\description{
fills missing countries in ILOSTAT data with 0
}
\examples{
\dontrun{
  a <- readSource("ILOSTAT", "EmplByActivityModelled", convert = TRUE)
}
}
\author{
Debbora Leip
}
