% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertIMF.R
\name{convertIMF}
\alias{convertIMF}
\title{convertIMF}
\usage{
convertIMF(x, subtype = "current_account")
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{A string}
}
\value{
magpie object with a completed dataset.
}
\description{
Converts data from IMF
}
\seealso{
\code{\link{convertIMF}}
}
