% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertISIMIP.R
\name{convertISIMIP}
\alias{convertISIMIP}
\title{convertISIMIP}
\usage{
convertISIMIP(x, subtype)
}
\arguments{
\item{x}{MAgPIE object on cellular level}

\item{subtype}{data subtype}
}
\value{
MAgPIE object on country level
}
\description{
convert data to ISO country level
}
\examples{
\dontrun{
a <- readSource("ISIMIP", convert = TRUE)
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Jan Philipp Dietrich, Felicitas Beier
}
