% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertJames2019.R
\name{convertJames2019}
\alias{convertJames2019}
\title{Convert updated James data}
\usage{
convertJames2019(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing James data region resolution}

\item{subtype}{subtype of GDP indicator to be selected}
}
\value{
GDP per capita in USD05 in PPP or MER as magpie object
}
\description{
Convert updated James data
}
\examples{

\dontrun{ a <- convertSource("James2019","IHME_USD05_PPP_pc")
}
}
\seealso{
\code{\link{readJames}}
}
\author{
David CHen
}
