% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertKoeppen.R
\name{convertKoeppen}
\alias{convertKoeppen}
\title{convertKoeppen}
\usage{
convertKoeppen(x, subtype = "iso")
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{Switch between different levels}
}
\value{
List of magpie objects with results on country level
}
\description{
Convert Koeppen climate zones on iso-country level
}
\examples{
\dontrun{
readSource("Koeppen", subtype = "iso", convert = TRUE)
}
}
\author{
Kristine Karstens
}
