% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertLPJmL.R
\name{convertLPJmL}
\alias{convertLPJmL}
\title{convertLPJmL}
\usage{
convertLPJmL(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Convert LPJmL content
}
\examples{

\dontrun{
  readSource("LPJmL", subtype="soilc", convert=TRUE)
}

}
\seealso{
\code{\link[=readLPJmL]{readLPJmL()}}
}
\author{
Kristine Karstens
}
