% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertMAgPIE.R
\name{convertMAgPIE}
\alias{convertMAgPIE}
\title{Convert MAgPIE data}
\usage{
convertMAgPIE(x, subtype)
}
\arguments{
\item{x}{input MAgPIE data on region-level}

\item{subtype}{Either "EmiAirPoll", "macBase", "co2tax", "supplyCurve_magpie_40", "abatparam_co2"}
}
\value{
magpie object
}
\description{
Convert MAgPIE data to country-level.
}
\author{
David Klein, Felix Schreyer
}
