% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPBL_MACC_2019.R
\name{convertPBL_MACC_2019}
\alias{convertPBL_MACC_2019}
\title{Convert subtypes of the PBL_MACC_2019 data}
\usage{
convertPBL_MACC_2019(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing PBL_MACC_2019 data on region level}

\item{subtype}{data subtype.
"ch4coal","ch4oil","ch4gas","ch4wstl","ch4wsts","ch4rice","ch4animals","ch4anmlwst",
"n2otrans","n2oadac","n2onitac","n2ofert","n2oanwst","n2owaste",
"HFC_tot", "SF6_tot", "PFC_tot" or "baseline_sources"}
}
\value{
PBL_MACC_2019 data as MAgPIE object for all subtypes aggregated to
country level
}
\description{
Convert subtypes from PBL_MACC_2019 to data on ISO country level.#'
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Florian Humpenoeder
}
