% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPBL_MACC_SSP2_2022.R
\name{convertPBL_MACC_SSP2_2022}
\alias{convertPBL_MACC_SSP2_2022}
\title{Convert subtypes of the PBL_MACC_SSP2_2022 data}
\usage{
convertPBL_MACC_SSP2_2022(x)
}
\arguments{
\item{x}{MAgPIE object containing PBL_MACC_SSP2_2022 data on region level}
}
\value{
PBL_MACC_SSP2_2022 data as MAgPIE object for all subtypes aggregated to
country level
}
\description{
Convert subtypes from PBL_MACC_SSP2_2022 to data on ISO country level.
}
\author{
Gabriel Abrahão
}
