% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertPEAP.R
\name{convertPEAP}
\alias{convertPEAP}
\title{Convert Population Estimates And Projections from the World Bank}
\usage{
convertPEAP(x)
}
\arguments{
\item{x}{MAgPIE object}
}
\value{
MAgPIE object
}
\description{
Convert data from the World Bank's Population Estimates And Projections
}
\author{
Johannes Koch
}
