% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertSoilGrids.R
\name{convertSoilGrids}
\alias{convertSoilGrids}
\title{convertSoilGrids}
\usage{
convertSoilGrids(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Convert SoilGrids content
}
\examples{

\dontrun{
  readSource("SoilGrids", subtype="cstock_0_30", convert=TRUE)
}

}
\seealso{
\code{\link{readSoilGrids}}
}
\author{
Kristine Karstens
}
