% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertVittis.R
\name{convertVittis}
\alias{convertVittis}
\title{convertVittis}
\usage{
convertVittis(x)
}
\arguments{
\item{x}{MAgPIE object to be converted}
}
\value{
A MAgPIE object containing national-scale costs of production for 
10 crops, disaggregated in 9 distinct cost elements
}
\description{
Aggregate production costs from Vittis dataset to MAgPIE crop 
categories and change unit from US$2000 to US$2005.
}
\author{
Debbora Leip
}
