% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctAndrijevic2019.R
\name{correctAndrijevic2019}
\alias{correctAndrijevic2019}
\title{correctAndrijevic2019}
\usage{
correctAndrijevic2019(x, subtype)
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{data to be returned:
"historical" for observed data until 2015
"projected" for projected SSP scenario data from 2015 to 2099}
}
\value{
corrected magpie object iso-level
}
\description{
correct Andrijevic governance data
}
\seealso{
\code{\link{readAndrijevic2019}}
}
\author{
Felicitas Beier
}
