% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctEDGARfood.R
\name{correctEDGARfood}
\alias{correctEDGARfood}
\title{correctEDGARfood}
\usage{
correctEDGARfood(x, subtype)
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{Type of data that should be read
\itemize{
\item \code{foodSystemEmi}: Total food system emissions of different countries
\item \code{foodSystemShare}: Share of food system emissions in total emissions
\item \code{foodSystemSector}: Food system emissions separated by country, sector and substance
}}
}
\value{
corrected magpie object
}
\description{
correct Edgar-food data
}
\seealso{
\code{\link{readEDGARfood}}
}
\author{
David Hoetten
}
