% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFAO_online.R
\name{correctFAO_online}
\alias{correctFAO_online}
\title{correctFAO_online}
\usage{
correctFAO_online(x, subtype)
}
\arguments{
\item{x}{MAgPIE object containing original values}

\item{subtype}{The FAO file type, e.g.: CBCrop}
}
\value{
Data as MAgPIE object
}
\description{
Corrects FAO data for known mismatches or insufficiencies
}
\examples{
\dontrun{
a <- readSource("FAO_online", "Crop", convert = TRUE)
}

}
\seealso{
\code{\link[=readFAO]{readFAO()}}, \code{\link[=readSource]{readSource()}},
}
\author{
Kristine Karstens
}
