% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctFoodSystemsDashboard.R
\name{correctFoodSystemsDashboard}
\alias{correctFoodSystemsDashboard}
\title{correctFoodSystemsDashboard}
\usage{
correctFoodSystemsDashboard(x, subtype)
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{Switch between different levels}
}
\value{
List of magpie objects
}
\description{
Correct FoodSystemsDashboard values that are questionable
}
\author{
David Chen
}
