% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctGAEZv4.R
\name{correctGAEZv4}
\alias{correctGAEZv4}
\title{correctGAEZv4}
\usage{
correctGAEZv4(x)
}
\arguments{
\item{x}{MAgPIE object provided by readGAEZv4 function}
}
\value{
MAgPIE object at 0.5 cellular level
}
\description{
Correct Global Agro-ecological Zones (GAEZ) data
}
\examples{
\dontrun{
readSource("GAEZv4", convert = "onlycorrect")
}

}
\author{
Felicitas Beier
}
