% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctILOSTAT.R
\name{correctILOSTAT}
\alias{correctILOSTAT}
\title{correctILOSTAT}
\usage{
correctILOSTAT(x, subtype)
}
\arguments{
\item{x}{unconverted magpie object from read-script}

\item{subtype}{Type of ILOSTAT data that should be read}
}
\value{
Data as MAgPIE object with corrected currency units
}
\description{
changes currency unit where applicable
}
\examples{
\dontrun{
  a <- readSource("ILOSTAT", "EmplByActivityModelled", convert = "onlycorrect")
}
}
\author{
Debbora Leip
}
