% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctKoeppen.R
\name{correctKoeppen}
\alias{correctKoeppen}
\title{correctKoeppen}
\usage{
correctKoeppen(x, subtype = "iso")
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{Switch between different levels}
}
\value{
List of magpie objects with results on cellular level
}
\description{
Correct Koeppen climate zones on cellular level
}
\examples{

\dontrun{
  readSource("Koeppen", subtype="cellular", convert="onlycorrect")
}
}
\author{
Kristine Karstens
}
