% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLandInG.R
\name{correctLandInG}
\alias{correctLandInG}
\title{correctLandInG}
\usage{
correctLandInG(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
corrected magpie object
}
\description{
correct LandInG data. Convert unit from ha to mio ha
}
\examples{
\dontrun{
a <- readSource("LandInG", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readLandInG}}
}
\author{
David Hoetten, Felicitas Beier
}
