% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmL.R
\name{correctLPJmL}
\alias{correctLPJmL}
\title{correctLPJmL}
\usage{
correctLPJmL(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct LPJmL content
}
\examples{
\dontrun{
readSource("LPJmL", subtype="soilc", convert="onlycorrect")
}

}
\seealso{
\code{\link{correctLPJmL}}
}
\author{
Kristine Karstens, Felicitas Beier
}
