% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmL_new.R
\name{correctLPJmL_new}
\alias{correctLPJmL_new}
\title{correctLPJmL_new}
\usage{
correctLPJmL_new(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
Convert LPJmL content (dummy function)
}
\examples{

\dontrun{
  readSource("LPJmL", convert="onlycorrect")
}

}
\seealso{
\code{\link[=readLPJmL]{readLPJmL()}}
}
\author{
Kristine Karstens
}
