% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmLClimateInput.R
\name{correctLPJmLClimateInput}
\alias{correctLPJmLClimateInput}
\title{correctLPJmLClimateInput}
\usage{
correctLPJmLClimateInput(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
Magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct LPJmL climate input variables
}
\examples{

\dontrun{
readSource("LPJmLClimateInput", subtype, convert="onlycorrect")
}

}
\seealso{
\code{\link{readLPJmLClimateInput}}
}
\author{
Marcos Alves, Felicitas Beier
}
