% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLUH2v2.R
\name{correctLUH2v2}
\alias{correctLUH2v2}
\title{correctLUH2v2}
\usage{
correctLUH2v2(x, subtype)
}
\arguments{
\item{x}{magpie object provided by the read function}

\item{subtype}{switch between different inputs}
}
\value{
List of magpie object with results on cellular level
}
\description{
Correct LUH2v2 content
}
\author{
Florian Humpenoeder, Stephen Wirth, Kristine Karstens, Felicitas Beier, Jan Philipp Dietrich,
 Edna J. Molina Bacca
}
