% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctMehta2022.R
\name{correctMehta2022}
\alias{correctMehta2022}
\title{correctMehta2022}
\usage{
correctMehta2022(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object in cellular resolution
}
\description{
correct lobal Area Equipped for Irrigation Dataset 1900-2015
from Mehta et al., 2022
}
\examples{

\dontrun{
  readSource("Mehta2022", convert="onlycorrect")
}
}
\author{
Felicitas Beier
}
