% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSiebert.R
\name{correctSiebert}
\alias{correctSiebert}
\title{correctSiebert}
\usage{
correctSiebert(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object in cellular resolution
}
\description{
correctc area equipped for irrigation for the period 1997-2002 from Siebert et al.; 2007 ;
             Global Map of Irrigation Areas version 4.0.1
}
\examples{

\dontrun{
  readSource("Siebert", convert="onlycorrect")
}
}
\author{
David Chen
}
