% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctSoilGrids.R
\name{correctSoilGrids}
\alias{correctSoilGrids}
\title{correctSoilGrids}
\usage{
correctSoilGrids(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Correct SoilGrids content
}
\examples{

\dontrun{
  readSource("SoilGrids", subtype="cstock_0_30", convert="onlycorrect")
}

}
\seealso{
\code{\link{readSoilGrids}}
}
\author{
Kristine Karstens
}
