% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctTNC2019.R
\name{correctTNC2019}
\alias{correctTNC2019}
\title{correctTNC2019}
\usage{
correctTNC2019(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct biome data
}
\examples{
\dontrun{
readSource("TNC2019", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readTNC2019}}
}
\author{
Patrick v. Jeetze
}
