% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctVittis.R
\name{correctVittis}
\alias{correctVittis}
\title{correctVittis}
\usage{
correctVittis(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
corrected mapie object on national-scale costs of production for 10 
crops, disaggregated in 9 distinct cost elements
}
\description{
Correct inconsistencies in crop naming in the Vittis dataset
}
\seealso{
\code{\link{readVittis}}
}
\author{
Debbora Leip
}
