% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctZabel2014.R
\name{correctZabel2014}
\alias{correctZabel2014}
\title{correctZabel2014}
\usage{
correctZabel2014(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\value{
magpie object on cellular level
}
\description{
correct Zabel crop suitability data
}
\examples{
\dontrun{
readSource("Zabel2014", convert = "onlycorrect")
}

}
\seealso{
\code{\link{readZabel2014}}
}
\author{
Patrick v. Jeetze
}
