% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadAndrijevic2019.R
\name{downloadAndrijevic2019}
\alias{downloadAndrijevic2019}
\title{downloadAndrijevic2019}
\usage{
downloadAndrijevic2019(subtype = "governance_obs_project")
}
\arguments{
\item{subtype}{file that should be downloaded from governance2019 repository}
}
\description{
This function downloads the governance indicator data available
             at https://github.com/marina-andrijevic/governance2019
}
\examples{
\dontrun{
downloadSource("Andrijevic2019", subtype = "governance_obs_project")
}

}
\seealso{
\code{\link{readAndrijevic2019}}
}
\author{
Felicitas Beier
}
