% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEDGAR6.R
\name{downloadEDGAR6}
\alias{downloadEDGAR6}
\title{downloadEDGAR}
\usage{
downloadEDGAR6(subtype = "n2o")
}
\arguments{
\item{subtype}{type in gas for receiving the newest data. type in specific filename for old version 4.31 data}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
download EDGAR emission data
}
\author{
Benjamin Leon Bodirsky,
}
