% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadEurostatLivestock.R
\name{downloadEurostatLivestock}
\alias{downloadEurostatLivestock}
\title{Download Eurostat Livestock Data}
\usage{
downloadEurostatLivestock(subtype)
}
\arguments{
\item{subtype}{Type of Eurostat Livestock data that should be read.}
}
\description{
Downloads the latest data and meta data form the Eurostat website.
}
