% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGFED.R
\name{downloadGFED}
\alias{downloadGFED}
\title{Download GFED}
\usage{
downloadGFED()
}
\description{
Download the GFED (Global Fire Emissions Database) dataset for all years not labeled
as beta, in addition to the emission factors.
}
\examples{
\dontrun{
a <- downloadSource()
}

}
\seealso{
\code{\link[madrat:downloadSource]{madrat::downloadSource()}} \code{\link[=readGFED]{readGFED()}}
}
\author{
Michael S. Crawford
}
