% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGGCMICropCalendar.R
\name{downloadGGCMICropCalendar}
\alias{downloadGGCMICropCalendar}
\title{downloadGGCMICropCalendar}
\usage{
downloadGGCMICropCalendar()
}
\description{
Download ISIMIP GGCMI crop calendar information and harvest area masks
}
\author{
David M Chen, Edna Molina Bacca
}
