% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadGGCMIHarvestAreaMask.r
\name{downloadGGCMIHarvestAreaMask}
\alias{downloadGGCMIHarvestAreaMask}
\title{downloadGGCMIHarvestArea
Download ISIMIP GGCMI harvested area mask information for rice 1 and rice 2, as well as spring/winter wheat

Downloads harvested area}
\usage{
downloadGGCMIHarvestAreaMask()
}
\description{
downloadGGCMIHarvestArea
Download ISIMIP GGCMI harvested area mask information for rice 1 and rice 2, as well as spring/winter wheat

Downloads harvested area
}
\author{
David M Chen
}
