% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadILOSTAT.R
\name{downloadILOSTAT}
\alias{downloadILOSTAT}
\title{downloadILOSTAT}
\usage{
downloadILOSTAT(subtype)
}
\arguments{
\item{subtype}{Type of ILOSTAT data that should be downloaded}
}
\value{
metadata entry
}
\description{
Download data from ILOSTAT
}
\examples{
\dontrun{
  downloadSource("ILOSTAT", "EmplByActivityModelled")
}
}
\author{
Debbora Leip
}
