% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadIPCCClimate.R
\name{downloadIPCCClimate}
\alias{downloadIPCCClimate}
\title{downloadIPCCClimate}
\usage{
downloadIPCCClimate()
}
\value{
Meta information on downloaded data
}
\description{
Download IPCC climate classification
}
\examples{

\dontrun{
readSource("IPCCClimate", convert="onlycorrect")
}

}
\author{
Kristine Karstens
}
