% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMehta2022.R
\name{downloadMehta2022}
\alias{downloadMehta2022}
\title{downloadMehta2022}
\usage{
downloadMehta2022()
}
\description{
download Global Area Equipped for Irrigation Dataset 1900-2015 from Mehta et al. (2022)
}
\examples{
\dontrun{
a <- downloadSource()
}

}
\seealso{
\code{\link[=downloadSource]{downloadSource()}} \code{\link[=readMehta2022]{readMehta2022()}}
}
\author{
Felicitas Beier
}
