% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadMissingIslands.R
\name{downloadMissingIslands}
\alias{downloadMissingIslands}
\title{downloadMissingIslands}
\usage{
downloadMissingIslands(subtype)
}
\arguments{
\item{subtype}{pop for population or gdp for GDP}
}
\description{
Download MissingIslands data (on population or GDP) for filling in data gaps from the
full datasets.
}
\examples{

\dontrun{ a <- downloadSource(type="MissingIslands", subtype = "pop")
}

}
\seealso{
\code{\link{downloadSource}}
}
\author{
Debbora Leip, Michael Crawford
}
