% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadNitrogenBoundariedGridded.R
\name{downloadNitrogenBoundariesGridded}
\alias{downloadNitrogenBoundariesGridded}
\title{downloadNitrogenBoundariesGridded}
\usage{
downloadNitrogenBoundariesGridded()
}
\description{
Download he grid-level regional nitrogen boundary datasets from Schulte-Uebbing et al. (2022)
}
\examples{
\dontrun{
 a <- downloadSource("NitrogenBoundariesGridded")
}
}
\seealso{
\code{\link[madrat:downloadSource]{madrat::downloadSource()}}
}
\author{
Michael S. Crawford
}
