% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTFP_USDA.R
\name{downloadTFP_USDA}
\alias{downloadTFP_USDA}
\title{downloadTFP_USDA}
\usage{
downloadTFP_USDA()
}
\value{
raw TFP file from USDA
}
\description{
Downloads data of input shares based on a TFP assessment from USDA.
}
\examples{

\dontrun{
a <- download("TFP_USD")
}

}
\seealso{
\code{\link{downloadSource}}
}
\author{
Edna J. Molina Bacca
}
