% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTFPUSDA.R
\name{downloadTFPUSDA}
\alias{downloadTFPUSDA}
\title{downloadTFPUSDA}
\usage{
downloadTFPUSDA()
}
\value{
raw TFP file from USDA
}
\description{
Downloads data of input shares based on a TFP assessment from USDA.
}
\examples{
\dontrun{
a <- download("TFP_USD")
}

}
\seealso{
\code{\link[=downloadSource]{downloadSource()}}
}
\author{
Edna J. Molina Bacca
}
