% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadTravelTimeNelson2019.R
\name{downloadTravelTimeNelson2019}
\alias{downloadTravelTimeNelson2019}
\title{downloadTravelTimeNelson2019}
\usage{
downloadTravelTimeNelson2019()
}
\description{
download Nelson 2019 paper
}
\author{
David M Chen
}
