% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downloadUN_PopDiv.R
\name{downloadUN_PopDiv}
\alias{downloadUN_PopDiv}
\title{Download UN_PopDiv}
\usage{
downloadUN_PopDiv()
}
\description{
Download UN_PopDiv dataset (World Population Prospects)
}
\examples{

\dontrun{ a <- downloadSource()
}
}
\seealso{
\code{\link{downloadSource}} \code{\link{readUN_PopDiv}}
}
\author{
Michael Crawford, Debbora Leip
}
