% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAdjustGrassi2021.R
\name{readAdjustGrassi2021}
\alias{readAdjustGrassi2021}
\title{readAdjustGrassi2021}
\usage{
readAdjustGrassi2021(subtype)
}
\arguments{
\item{subtype}{Either "data" or "weight"}
}
\value{
Adjustement factors or weight as MAgPIE object
}
\description{
Read in data from Grassi et al. 2021.
Adjustement factors for emission in GtCO2 yr-1.
}
\examples{
\dontrun{
readSource("AdjustGrassi2021", subtype = "data")
}
}
\author{
Michael Windisch, Florian Humpenoeder
}
