% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readAndrijevic2019.R
\name{readAndrijevic2019}
\alias{readAndrijevic2019}
\title{readAndrijevic2019}
\usage{
readAndrijevic2019(subtype)
}
\arguments{
\item{subtype}{data used from governance2019 repository}
}
\value{
governance index data at iso-country level
}
\description{
read in governance index data from Andrijevic et al. 2019
}
\examples{
\dontrun{
a <- readSource("Andrijevic2019")
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Felicitas Beier
}
