% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCarbonLTS.R
\name{readCarbonLTS}
\alias{readCarbonLTS}
\title{readCarbonLTS}
\usage{
readCarbonLTS(subtype)
}
\arguments{
\item{subtype}{data subtype.}
}
\value{
magpie object of the FRA 2015 data
}
\description{
Read-in an Long term carbon storage data for historical period.
}
\examples{
\dontrun{
a <- readSource("CarbonLTS", "Lauk_et_al")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Abhijeet Mishra
}
