% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCEDS2021.R
\name{readCEDS2021}
\alias{readCEDS2021}
\title{readCEDS2021}
\usage{
readCEDS2021()
}
\value{
MAgPIE object
}
\description{
reads in emission data from the CEDS database
}
\author{
Benjamin Leon Bodirsky, David Klein
}
