% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCEDS2025.R
\name{readCEDS2025}
\alias{readCEDS2025}
\title{readCEDS2025}
\usage{
readCEDS2025()
}
\value{
MAgPIE object
}
\description{
reads in emission data from the CEDS database
publication: https://zenodo.org/records/15059443
}
\author{
Pascal Weigmann
}
