% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCEDS_CMIP7.R
\name{readCEDS_CMIP7}
\alias{readCEDS_CMIP7}
\title{readCEDS_CMIP7}
\usage{
readCEDS_CMIP7()
}
\value{
MAgPIE object
}
\description{
reads in harmonized emission data used for CMIP7 from three files.
country_file for energy related emissions on country level and global bunker
emissions, global_file for F-Gases and AFOLU on global level,
cmip7-historical-ghg-eq for GHG totals.
Source: https://zenodo.org/records/15059443
}
\author{
Pascal Weigmann
}
