% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readDomingues.R
\name{readDomingues}
\alias{readDomingues}
\title{readDomingues}
\usage{
readDomingues()
}
\value{
Data as MAgPIE object with average yields and costs of betr and begr
}
\description{
Reads biomass yields (GJ/ha/year) and factor costs (US$ 2018/GJ)
for bioenergy crops from Domingues et al., 2022 (DOI: 10.5281/zenodo.6913709)
}
\author{
Eva Bleidorn
}
