% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR6.R
\name{readEDGAR6}
\alias{readEDGAR6}
\title{readEDGAR6}
\usage{
readEDGAR6(subtype)
}
\arguments{
\item{subtype}{type in gas}
}
\value{
List of magpie objects with results on country level, weight, unit and description.
}
\description{
download EDGAR 5 and 6 emission data
}
\author{
Benjamin Leon Bodirsky,
}
