% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR8.R
\name{readEDGAR8}
\alias{readEDGAR8}
\title{readEDGAR8}
\usage{
readEDGAR8()
}
\value{
magpie objects of EDGAR historical emissions (MtCO2eq)
}
\description{
read EDGAR 8 emission data
}
\seealso{
\code{\link{readSource}}
}
\author{
Pascal Weigmann
}
