% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR_GHG.R
\name{readEDGAR_GHG}
\alias{readEDGAR_GHG}
\title{readEDGAR_GHG}
\usage{
readEDGAR_GHG()
}
\value{
magpie objects of EDGAR historical emissions (MtCO2eq)
}
\description{
read EDGAR GHG emission data
}
\seealso{
\code{\link{readSource}}
}
\author{
Pascal Weigmann
}
