% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGAR_LU.R
\name{readEDGAR_LU}
\alias{readEDGAR_LU}
\title{Read EDGAR_LU}
\usage{
readEDGAR_LU(subtype = "CO2")
}
\arguments{
\item{subtype}{emission type}
}
\value{
magpie object EDGAR_LU data
}
\description{
Read-in EDGAR_LU csv file as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "EDGAR_LU")
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Florian Humpenoeder
}
