% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGARfood.R
\name{readEDGARfood}
\alias{readEDGARfood}
\title{readEDGARfood}
\usage{
readEDGARfood(subtype)
}
\arguments{
\item{subtype}{Type of data that should be read
\itemize{
\item \code{foodSystemEmi}: Total food system emissions of different countries
\item \code{foodSystemShare}: Share of food system emissions in total emissions
\item \code{foodSystemSector}: Food system emissions separated by country, sector and substance
}}
}
\value{
A magpie object with foodsystem total emissions,
emission shares or sector and substance specific emission, depending on subtype
}
\description{
read Edgar-food v6 food system emission data
}
\author{
David Hötten
}
