% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostatLivestock.R
\name{readEurostatLivestock}
\alias{readEurostatLivestock}
\title{Read EUROSTAT_livestock}
\usage{
readEurostatLivestock(subtype)
}
\arguments{
\item{subtype}{Type of FAO data that should be read. Available types are:
\itemize{
\item    \code{EggPopProd} : Production of eggs for consumption and number of laying hens
\item    \code{LayingHensPop} : Laying hens population - annual data
\item    \code{PoultryPop} : Poultry - annual data
\item    \code{PigPop} : Pig population - annual data
\item    \code{BovinePop} : Bovine population - annual data
\item    \code{Nuts2Pop} : Animal populations by NUTS 2 regions
\item    \code{MeatProd} : Meat production and foreign trade - head - monthly data
\item    \code{MilkNuts2Prod} : Production of cow's milk on farms by NUTS 2 regions
\item    \code{MilkProd} : Production and utilization of milk on the farm - annual data
}}
}
\value{
EUROSTAT livestock data as MAgPIE object
}
\description{
Read in EUROSTAT livestock data
}
\examples{
\dontrun{
a <- readSource("EurostatLivestock", "MeatProd")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
David Hötten
}
