% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExpertGuessSSPLivestockProduction.R
\name{readExpertGuessSSPLivestockProduction}
\alias{readExpertGuessSSPLivestockProduction}
\title{readExpertGuessSSPLivestockProduction}
\usage{
readExpertGuessSSPLivestockProduction(subtype)
}
\arguments{
\item{subtype}{: Available subtypes are the combination of data source and ssps 1 to 5. E.g. "default:ssp1"}
}
\value{
magpie object containing the expert guesses
}
\description{
{Read the Expert Guesses for future Livestock Production for the SSP Scenarios
}
}
\examples{
\dontrun{ 
a <- readSource("ExperGuessSSPLivestockProduction", "ssp1")
}
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Stephen Wirth, Marcos Alves
}
