% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFAO_TradeMatrix.R
\name{readFAO_TradeMatrix}
\alias{readFAO_TradeMatrix}
\title{Read FAO_TradeMatrix}
\usage{
readFAO_TradeMatrix(subtype)
}
\arguments{
\item{subtype}{subsets of the detailed trade matrix to read in. Very large csv needs to be read in chunks
separated by export/import quantities and values, as well as kcr, kli and kothers (not in kcr nor kli)
Options are all combinations of c("import_value", "import_qty", "export_value", "export_quantity" X c("kcr", "kli", "kothers"))
import is import side reporting while export is export-sde reporting}
}
\value{
FAO data as MAgPIE object
}
\description{
Read in FAOSTAT detail trade matrix.
FAOSTAT does not balance or harmonize the import/export side reporting, effectively creating 2 slightly different datasets.
Furthermore, in terms of trade value, exporters are "usuallY" reporting FOB, while importers report CIF. Difference in value,
given identical qty, is thus the transport margin.
}
\examples{
\dontrun{
a <- readSource("FAO_TradeMatrix", "import_value")
}
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
David C
}
