% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFishstatJ_FAO.R
\name{readFishstatJ_FAO}
\alias{readFishstatJ_FAO}
\title{readFishstatJ_FAO}
\usage{
readFishstatJ_FAO(subtype = "Production")
}
\arguments{
\item{subtype}{data subtype needed. Either "exportsValue", "exportsQuantity", or "Production"}
}
\value{
magpie object of either tonnes of liveweight or 1000 current USD
}
\description{
Reads data of fisheries generated using the FishstatJ app of FAO.
Read-in specifically, exports_value, exports_quantity, and/or overall production of fish/aquatic products.
}
\examples{
\dontrun{
a <- readSource("FishstatJ_FAO", "Production")
a <- readSource("FishstatJ_FAO", "exportsQuantity")
a <- readSource("FishstatJ_FAO", "exportsValue")
}

}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Edna J. Molina Bacca
}
