% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFoodSystemsDashboard.R
\name{readFoodSystemsDashboard}
\alias{readFoodSystemsDashboard}
\title{readFoodSystemsDashboard}
\usage{
readFoodSystemsDashboard(subtype)
}
\arguments{
\item{subtype}{Currently either "Processed" for Processed food expenditures per capita or
"Industrial Processing Share"}
}
\value{
governance index data at iso-country level
}
\description{
read in Food Systems Dashboard indicators
}
\examples{
\dontrun{
a <- readSource("FoodSystemsDashboard")
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
David Chen
}
