% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGAEZv4.R
\name{readGAEZv4}
\alias{readGAEZv4}
\title{readGAEZv4}
\usage{
readGAEZv4(subtype = "MCzones")
}
\arguments{
\item{subtype}{Subtype to be read}
}
\value{
MAgPIE object at 0.5 cellular level
}
\description{
Read in data from the Global Agro-ecological Zones (GAEZ) data set version 4
}
\examples{
\dontrun{
readSource("GAEZv4", convert = "onlycorrect")
}

}
\author{
Felicitas Beier
}
