% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGFED.R
\name{readGFED}
\alias{readGFED}
\title{Read GFED}
\usage{
readGFED()
}
\value{
A MAgPIE object with the GFED emissions data and sub-dimensions Partition and Species.
}
\description{
Read the Global Fire Emissions Database over the available years and store
them in a MAgPIE object with two sub-dimensions: Species (e.g. DM, CO2, CH4, or N2O),
and Partition (e.g. SAVA, BORF, TEMF, DEFO, PEAT, AGRI). Emissions are reported in Mt X.
For more information, see: https://globalfiredata.org/pages/data/. Due to runtime considerations,
this function only calculates a small sample of the total emissions (CH4, N2O, NOx, NH3) and
burning partitions (AGRI), which are currently being used for validation.
}
\examples{
\dontrun{
a <- readSource("GFED")
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Michael S. Crawford
}
