% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGGCMICropCalendar.R
\name{readGGCMICropCalendar}
\alias{readGGCMICropCalendar}
\title{readGGCMICropCalendar}
\usage{
readGGCMICropCalendar()
}
\value{
A MAgPIE 0.5deg resolution, day of the year
}
\description{
Reads in planting, maturity, and harvest date based on GGCMI hybrid product of crop calendars
}
\author{
David Chen
}
