% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGTAP81.R
\name{readGTAP81}
\alias{readGTAP81}
\title{readGTAP81}
\usage{
readGTAP81(subtype)
}
\arguments{
\item{subtype}{GTAP header that should be read. Available headers are listed in the contents_xxx.csv in the
GTAP81 source folder for each GTAP file (BaseData.har, BaseRate.har, BaseView.har, CO2.har, gsdvole.har,
GTAPSam.har, Sets.har, TStrade.har, and Default.prm - file is determined based on the header via a mapping
and does not need to be specified).}
}
\value{
GTAP data as MAgPIE object
}
\description{
Read in data from GTAP 8.1 baseyears 2007 and 2004. All values in mio. current US$MER.
}
\examples{
\dontrun{
readSource("GTAP81", subtype = "SF01")
}
}
\author{
Debbora Leip
}
