% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGTAPv8v9.R
\name{readGTAPv8v9}
\alias{readGTAPv8v9}
\title{readGTAPv8v9}
\usage{
readGTAPv8v9(subtype)
}
\arguments{
\item{subtype}{GTAP version and header that should be read. Available versions are "81" and "9"
Available headers are listed in the contents_xxx.csv in the
GTAPv8v9 source folder for each GTAP file (BaseData.har, BaseRate.har, BaseView.har, CO2.har, gsdvole.har,
GTAPSam.har, Sets.har, TStrade.har, and Default.prm - file is determined based on the header via a mapping
and does not need to be specified).}
}
\value{
GTAP data as MAgPIE object
}
\description{
Read in data from GTAP 8.1 baseyears 2007 and 2004. All values in mio. current US$MER.
}
\examples{
\dontrun{
readSource("GTAPv8v9", subtype = "81:SF01")
}
}
\author{
Debbora Leip, David M. Chen
}
