% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHalfEarth.R
\name{readHalfEarth}
\alias{readHalfEarth}
\title{readHalfEarth}
\usage{
readHalfEarth(subtype = "GLOBIO4")
}
\arguments{
\item{subtype}{Data source to be read from}
}
\value{
MAgPIE object containing biodiveristy protection area at cellular level
}
\description{
Read in Half Earth data set containing conservation area
for biodiversity protection based on the Half-Earth approach
}
\examples{
\dontrun{
readSource("HalfEarth", subtype = "GLOBIO4", convert = "onlycorrect")
}

}
\author{
Felicitas Beier
}
