% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHI.R
\name{readHI}
\alias{readHI}
\title{Read allometric coeffizients for residue to harvest translation}
\usage{
readHI()
}
\value{
magpie object with the dimension crops and coeffizients
}
\description{
Read-in a file containing the allometric coeffizients of MAgPIE crop types.
Values are assembled from various literature sources, and the weighting and allocation is done in the
spreadsheet crop_specifications_*.ods for different versions of the file
}
\examples{
\dontrun{
a <- readSource("HI")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Benjamin Leon Bodrisky, Kristine Karstens
}
