% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHistBioplasticProd.R
\name{readHistBioplasticProd}
\alias{readHistBioplasticProd}
\title{readHistBioplasticProd}
\usage{
readHistBioplasticProd()
}
\value{
data as MAgPIE object
}
\description{
reads data on historic bioplastic production from csv
(source: "Biopolymers facts and statistics 2021" page 58, IfBB)
}
\examples{
\dontrun{
  readSource("HistBioplasticProd")
}
}
\author{
Debbora Leip
}
