% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHoulton2018.R
\name{readHoulton2018}
\alias{readHoulton2018}
\title{readHoulton2018}
\usage{
readHoulton2018()
}
\value{
magpie object global resolution by landuse types in Mt of N weathered from rocks each year
}
\description{
Read Houlton 2018 model on weathering of N from rocks
}
\examples{
\dontrun{
readSource("Houlton2018", convert = FALSE)
}
}
\author{
David M Chen
}
