% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHourlyLaborCostsChina.R
\name{readHourlyLaborCostsChina}
\alias{readHourlyLaborCostsChina}
\title{readHourlyLaborCostsChina}
\usage{
readHourlyLaborCostsChina()
}
\value{
data as MAgPIE object
}
\description{
Read in data on hourly labor costs in agriculture for China.
}
\examples{
\dontrun{
readSource("HourlyLaborCostsChina")
}
}
\author{
Debbora Leip
}
