% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA.R
\name{readIEA}
\alias{readIEA}
\title{Read IEA}
\usage{
readIEA(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "EnergyBalances", "EnergyBalances-2023", or "Emissions".
\itemize{
\item "EnergyBalances": IEA energy balances until 2020 (incomplete 2021), data updated in Aug 2022,
the current default for REMIND input data
\item "EnergyBalances-latest": IEA energy balances until 2021 (incomplete 2022), data updated in Aug 2023,
currently used for comparisons only
}}
}
\value{
magpie object of the IEA
}
\description{
Read-in an IEA csv file as magpie object
}
\examples{
\dontrun{
a <- readSource(type = "IEA", subtype = "EnergyBalances")
}

}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Anastasis Giannousakis, Lavinia Baumstark, Renato Rodrigues, Falk Benke
}
