% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA2.R
\name{readIEA2}
\alias{readIEA2}
\title{Read IEA}
\usage{
readIEA2(subtype)
}
\arguments{
\item{subtype}{data subtype. Either "EnergyBalances", "CHPreport" or "Emissions")}
}
\value{
magpie object of the IEA
}
\description{
Read-in an IEA csv file as magpie object
}
\examples{
\dontrun{
a <- readSource(type = "IEA", subtype = "EnergyBalances")
}

}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Anastasis Giannousakis, Lavinia Baumstark, Renato Rodrigues
}
