% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_EEI.R
\name{readIEA_EEI}
\alias{readIEA_EEI}
\title{Read-in data from IEA End Uses and Efficiency Indicators Database}
\usage{
readIEA_EEI()
}
\description{
Read-in data from IEA End Uses and Efficiency Indicators Database
}
\author{
Falk Benke
}
