% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIEA_EnergyEfficiencyIndicators.R
\name{readIEA_EnergyEfficiencyIndicators}
\alias{readIEA_EnergyEfficiencyIndicators}
\title{Read-in data from IEA End Uses and Efficiency Indicators Database}
\usage{
readIEA_EnergyEfficiencyIndicators()
}
\description{
Read-in data from IEA End Uses and Efficiency Indicators Database
}
\author{
Falk Benke
}
