% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readHistBioplasticProd.R
\name{readIfBB}
\alias{readIfBB}
\title{readIfBB}
\usage{
readIfBB()
}
\value{
data as MAgPIE object
}
\description{
reads data on historic bioplastic production from csv
source: "Biopolymers facts and statistics 2021" page 58, Institute for Bioplastics and Biocomposites (IfBB)
}
\examples{
\dontrun{
  readSource("HistBioplasticProd")
}
}
\author{
Debbora Leip
}
