% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIFPRIsubsidy.R
\name{readIFPRIsubsidy}
\alias{readIFPRIsubsidy}
\title{readIFPRIsubsidy}
\usage{
readIFPRIsubsidy()
}
\value{
magpie object of agricultural subsidies
}
\description{
read subsidies for crops and livestock (and non-allocated) from IFPRI table
}
\examples{
\dontrun{
  readSource("IFPRIsubsidy")
}
}
\author{
Debbora Leip
}
