% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readImageMacc.R
\name{readImageMacc}
\alias{readImageMacc}
\title{Read in ImageMacc Costcurves for different subtypes}
\usage{
readImageMacc(subtype)
}
\arguments{
\item{subtype}{data subtype.
Options are: "CH4_Energy_Industry", "CH4_Landuse",
"N2O_Energy_Industry", "N2O_Landuse",
"HFC_tot", "SF6_tot", "PFC_tot" or "baseline_sources"}
}
\value{
magpie object of the ImageMacc data
}
\description{
Read in ImageMacc Costcurves for different subtypes, which are based on a Paper
fro Lucas et al 2007 (http://linkinghub.elsevier.com/retrieve/pii/S1462901106001316)
Rows are removed, the dataframe is reshaped and numbers are replaced by descriptions.
}
\examples{
\dontrun{
a <- readSource("ImageMacc", "CH4_Energy_Industry")
a <- readSource("ImageMacc", "CH4_Landuse")
a <- readSource("ImageMacc", "N2O_Energy_Industry")
a <- readSource("ImageMacc", "N2O_Landuse")
a <- readSource("ImageMacc", "HFC_tot")
a <- readSource("ImageMacc", "SF6_tot")
a <- readSource("ImageMacc", "PFC_tot")
a <- readSource("ImageMacc", "baseline_sources")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Nele Steinmetz
}
