% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIPCC.R
\name{readIPCC}
\alias{readIPCC}
\title{Read in IPCC emissions}
\usage{
readIPCC(subtype)
}
\arguments{
\item{subtype}{data subtype}
}
\value{
magpie object of the IPCC data
}
\description{
Read in IPCC data: \itemize{ \item Read in IPCC emissions from livestock and
manure management. Source: IPCC Guidelines for National Greenhouse Gas
Inventories (2006); Chapter 10: Emissions from Livestock and Manure
Management.  \item Read in IPCC emissions from Lime and urea application.
Source: IPCC Guidelines for National Greenhouse Gas Inventories (2006);
Chapter 11: N2O Emissions from managed Soils and Co2 Emissions from Lime and
Urea Application.  \item Read in IPCC efficiency factors for burning of
residue. Source: IPCC Guidelines for Natinal Greenhouse Gas Inventories
(2006); Chapter 02: Generic Methodologies applicable to multiple Land-use
Categories. \item Read in soil related stock change factors for carbon and
manure parameterization. Source: IPCC Guidelines for National Greenhouse Gas
Inventories (2006); Chapter 5: Cropland.}
}
\examples{
\dontrun{
a <- readSource("IPCC", "awmsShr")
a <- readSource("IPCC", "awmsEfCh4")
a <- readSource("IPCC", "awmsParCh4")
a <- readSource("IPCC", "nExcrRate")
a <- readSource("IPCC", "awmsconfef3", convert = FALSE)
a <- readSource("IPCC", "fracgasms", convert = FALSE)
a <- readSource("IPCC", "fraclossms", convert = FALSE)
a <- readSource("IPCC", "emissionfactors", convert = FALSE)
a <- readSource("IPCC", "rescombusteff", convert = FALSE)
a <- readSource("IPCC", "efnsoil", convert = FALSE)
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Nele Steinmetz, Stephen Wirth, Jan Philipp Dietrich,
Kristine Karstens
}
