% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readIPCCClimate.R
\name{readIPCCClimate}
\alias{readIPCCClimate}
\title{readIPCCClimate}
\usage{
readIPCCClimate()
}
\value{
Magpie object with results on cellular level for 12 IPCC climate zone types
}
\description{
Read IPCC climate classification
}
\examples{
\dontrun{
readSource("IPCCClimate", convert = "onlycorrect")
}
}
\author{
Kristine Karstens
}
