% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKoeppen.R
\name{readKoeppen}
\alias{readKoeppen}
\title{readKoeppen}
\usage{
readKoeppen(subtype = "iso")
}
\arguments{
\item{subtype}{Switch between different levels}
}
\value{
List of magpie objects with results on country or cellular level
}
\description{
Read Koeppen climate zones on iso-country and cellular level
}
\examples{
\dontrun{
readSource("Koeppen", subtype = "iso")
}

}
\author{
Kristine Karstens
}
