% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLandInG.R
\name{readLandInG}
\alias{readLandInG}
\title{readLandInG}
\usage{
readLandInG(subtype = "physicalArea")
}
\arguments{
\item{subtype}{Type of LandInG data that should be read:
\itemize{
\item \code{physicalArea}: Cropland extend/ physical cropping area separated in irrigated and rainfed
\item \code{harvestedArea}: Harvested area separated in different crop types
}}
}
\value{
magpie object
}
\description{
Reads in LandInG data
}
\examples{
\dontrun{
A <- readSource("LandInG", subtype = "harvestedArea", aggregate = FALSE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Felicitas Beier
}
