% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLanduseToolbox.R
\name{readLanduseToolbox}
\alias{readLanduseToolbox}
\title{readLanduseToolbox}
\usage{
readLanduseToolbox(subtype = "physicalArea")
}
\arguments{
\item{subtype}{Type of toolbox output data that should be read:
\itemize{
\item \code{physicalArea}: Cropland extend/ physical cropping area separated in irrigated and rainfed
\item \code{harvestedArea}: Harvested area separated in different crop types
}}
}
\value{
magpie object containing output of the toolbox
}
\description{
Reads in landuse output of LPJmL input toolbox
}
\examples{
\dontrun{
A <- readSource("LanduseToolbox", subtype = "harvestedArea", aggregate = FALSE)
}

}
\seealso{
\code{\link{readSource}}
}
\author{
David Hoetten
}
