% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLi2020.R
\name{readLi2020}
\alias{readLi2020}
\title{readLi2020}
\usage{
readLi2020()
}
\value{
MAgPIE object on cellular level with yields for Eucalypt, Miscanthus, Poplar,
Switchgrass, Willow, Best_crop and Best_crop_type.
}
\description{
Read lignocellulosic bioenergy crop yield maps from Li et al. (2020),
doi:10.5194/essd-12-789-2020. Random-forest upscaling of field observations using
climate and soil conditions. Yields in t DM ha-1 yr-1, ~year 2010.
}
\examples{
\dontrun{
a <- readSource("Li2020")
}
}
\author{
Kristine Karstens
}
