% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmL.R
\name{readLPJmL}
\alias{readLPJmL}
\title{readLPJmL}
\usage{
readLPJmL(subtype = "LPJmL5:CRU4p02.soilc")
}
\arguments{
\item{subtype}{Switch between different input}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LPJmL content
}
\examples{

\dontrun{
readSource("LPJmL", subtype="LPJmL5:CRU4p02.soilc", convert="onlycorrect")
}

}
\seealso{
\code{\link[=readLPJ]{readLPJ()}}
}
\author{
Kristine Karstens, Abhijeet Mishra, Felicitas Beier
}
