% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmL_new.R
\name{readLPJmL_new}
\alias{readLPJmL_new}
\title{readLPJmL_new}
\usage{
readLPJmL_new(
  subtype = "LPJmL4_for_MAgPIE_84a69edd:GSWP3-W5E5:historical:soilc"
)
}
\arguments{
\item{subtype}{Switch between different input (eg. "LPJmL5.2_Pasture:IPSL_CM6A_LR:ssp126_co2_limN_00:soilc_past_hist")}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
Read LPJmL content
}
\examples{

\dontrun{
readSource("LPJmL_new", convert=FALSE)
}

}
\seealso{
\code{\link{readLPJ}}
}
\author{
Kristine Karstens, Abhijeet Mishra, Felicitas Beier, Marcos Alves
}
