% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmLClimateInput.R
\name{readLPJmLClimateInput}
\alias{readLPJmLClimateInput}
\title{readLPJmLClimateInput}
\usage{
readLPJmLClimateInput(
  subtype = "ISIMIP3bv2:MRI-ESM2-0:ssp370:temperature",
  subset = "annualMean"
)
}
\arguments{
\item{subtype}{Switch between different inputs,
e.g. "ISIMIP3bv2:MRI-ESM2-0:ssp370:1850-2014:tas"
Available variables are: * tas -
* wet -
* per -}

\item{subset}{Switch between different subsets of the same subtype
Available options are: "annualMean", "annualSum",
"monthlyMean", "monthlySum",
"wetDaysMonth"
Note that not all subtype-subset combinations make sense}
}
\value{
MAgPIE objects with results on cellular level.
}
\description{
Read Climate data used as LPJmL inputs into MAgPIE objects
}
\examples{
\dontrun{
readSource("LPJmLClimateInput", subtype, convert = "onlycorrect")
}

}
\seealso{
\code{\link{readLPJmLClimateInput}}
}
\author{
Marcos Alves, Kristine Karstens, Felicitas Beier
}
