% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLPJmLInputs.R
\name{readLPJmLInputs}
\alias{readLPJmLInputs}
\title{readLPJmLInputs}
\usage{
readLPJmLInputs(subtype = "lakeshare")
}
\arguments{
\item{subtype}{Switch between different inputs}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description.
}
\description{
This function reads in LPJmL inputs (inputs to LPJmL)
}
\examples{

\dontrun{ 
readSource("LPJmLInputs", subtype="lakeshare", convert=FALSE)
}

}
\author{
Felicitas Beier
}
