% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readLUH2v2.R
\name{readLUH2v2}
\alias{readLUH2v2}
\title{readLUH2v2}
\usage{
readLUH2v2(subtype)
}
\arguments{
\item{subtype}{switch between different inputs}
}
\value{
List of magpie objects with results on cellular level, weight, unit and description
}
\description{
read LUH inputs
}
\author{
Florian Humpenoeder, Stephen Wirth, Kristine Karstens, Felicitas Beier,
Jan Philipp Dietrich, Patrick v. Jeetze
}
