% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readMehta2022.R
\name{readMehta2022}
\alias{readMehta2022}
\title{readMehta2022}
\usage{
readMehta2022()
}
\description{
reads in Global Area Equipped for Irrigation for years 1900-2015 from Mehta et al. (2022)
}
\examples{

\dontrun{ a <- readSource("Mehta2022")
}
}
\seealso{
\code{\link[=correctMehta2022]{correctMehta2022()}}
}
\author{
Felicitas Beier
}
