% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readNitrogenBoundariesGridded.R
\name{readNitrogenBoundariesGridded}
\alias{readNitrogenBoundariesGridded}
\title{readNitrogenBoundariesGridded}
\usage{
readNitrogenBoundariesGridded()
}
\value{
A MAgPIE object with the critical nitrogen surplus in kg N per ha per yr in 2010
}
\description{
Read the grid-level regional nitrogen boundary datasets from Schulte-Uebbing et al. (2022).
For the moment, this only uses the critical N surplus in all agricultural land (arable + intensively
managed grassland) in view of all thresholds simultaneously, measured in kg N per ha per yr.
}
\examples{
\dontrun{
 a <- readSource("NitrogenBoundariesGridded")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Michael S. Crawford
}
