% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPBL_MACC_2022.R
\name{readPBL_MACC_2022}
\alias{readPBL_MACC_2022}
\title{Read in PBL MAC curves from Harmsen_et_al_2022 for different subtypes and subsets}
\usage{
readPBL_MACC_2022(subtype, subset)
}
\arguments{
\item{subtype}{data subtype.
"ch4coal","ch4oil","ch4gas","ch4wstl","ch4wsts","ch4rice","ch4animals","ch4anmlwst",
"n2otrans","n2oadac","n2onitac","n2ofert","n2oanwst","n2owaste"}

\item{subset}{data subset.
"Default", "Optimistic", "Pessimistic"}
}
\value{
magpie object of the PBL_MACC_2022 data
}
\description{
Read in PBL MAC curves from Harmsen_et_al_2022 for different subtypes and subsets
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Michael Windisch, Florian Humpenoeder
}
