% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPEAP.R
\name{readPEAP}
\alias{readPEAP}
\title{Read Population Estimates And Projections from the World Bank}
\usage{
readPEAP()
}
\value{
magpie object
}
\description{
Read-in xlsx file from the World Bank's Population Estimates And Projections
}
\author{
Johannes Koch
}
