% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readProtectArea.R
\name{readProtectArea}
\alias{readProtectArea}
\title{readProtectArea}
\usage{
readProtectArea()
}
\value{
List of magpie objects with results on cellular level
}
\description{
Read conservation priority areas (in Mha)
}
\examples{
\dontrun{
readSource("ProtectArea", convert = "onlycorrect")
}

}
\author{
David Chen, Felicitas Beier
}
