% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readPyrolysisConditions.R
\name{readPyrolysisConditions}
\alias{readPyrolysisConditions}
\title{readPyrolysisConditions}
\usage{
readPyrolysisConditions(subtype = "Schmidt_2019")
}
\arguments{
\item{subtype}{Available subtypes: "Schmidt_2019", "Woolf_2014",
"Buffi_2024" and "Cornelissen_2016"}
}
\value{
Magpie object with the dimensions "properties" and "process_cond"
}
\description{
Reads in parameter specifications for different pyrolysis process
conditions from several literature sources, which are available as separate
source data files.
}
\examples{
\dontrun{
a <- readSource("PyrolysisConditions")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Isabelle Weindl
}
