% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readSiebert.R
\name{readSiebert}
\alias{readSiebert}
\title{readSiebert}
\usage{
readSiebert()
}
\value{
magpie object in cellular resolution
}
\description{
Read area equipped for irrigation for the period 1997-2002 from Siebert et al.; 2007 ;
             Global Map of Irrigation Areas version 4.0.1
}
\examples{

\dontrun{
  readSource("Siebert", convert="onlycorrect")
}
}
\author{
David Chen
}
