% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTFP_USDA.R
\name{readTFP_USDA}
\alias{readTFP_USDA}
\title{readTFP_USDA}
\usage{
readTFP_USDA()
}
\value{
magpie object with fractions of different input factors in the overall production value
}
\description{
Reads the input shares from USDA's Agricultural total factor productivity growth indices assessment.
}
\examples{

\dontrun{
a <- readSource("TFP_USD")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Edna J. Molina Bacca
}
