% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTFPUSDA.R
\name{readTFPUSDA}
\alias{readTFPUSDA}
\title{readTFPUSDA}
\usage{
readTFPUSDA()
}
\value{
magpie object with fractions of different input factors in the overall production value
}
\description{
Reads the input shares from USDA's Agricultural total factor productivity growth indices assessment.
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Edna J. Molina Bacca
}
