% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTNC2019.R
\name{readTNC2019}
\alias{readTNC2019}
\title{readTNC2019}
\usage{
readTNC2019()
}
\value{
Returns magpie object with a share of each spatial unit belonging to a biogeographic realm and major habitat type.
}
\description{
Reads geospatial data from 'the nature conservancy' on ecoregions, major habitat types (MHT, or biome types) and biogeographic realms.
}
\examples{
\dontrun{
readSource("TNC2019", convert = "onlycorrect")
}
}
\author{
Patrick v. Jeetze
}
