% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readTransportCostsGTAP.R
\name{readTransportCostsGTAP}
\alias{readTransportCostsGTAP}
\title{Read Transport Costs}
\usage{
readTransportCostsGTAP()
}
\value{
Transport Costs in USD
}
\description{
Read in  Transport Costs from GTAP
}
\examples{
\dontrun{
a <- readSource(type = "TransportCostsGTAP")
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
David Chen
}
