% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readVittis.R
\name{readVittis}
\alias{readVittis}
\title{readVittis}
\usage{
readVittis()
}
\value{
National-scale costs of production for 10 crops, disaggregated in 9 distinct cost elements
}
\description{
Read-in cost data from Vittis dataset.
}
\examples{

\dontrun{ a <- readSource("Vittis")
}
}
\seealso{
\code{\link[=readSource]{readSource()}}
}
\author{
Debbora Leip
}
