% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readZhang2015.R
\name{readZhang2015}
\alias{readZhang2015}
\title{Read Nitrogen Use Efficiency Paramerst from Zhang et al 2015}
\usage{
readZhang2015()
}
\value{
MAgPIE object
}
\description{
Publication: Zhang, Xin, Eric A. Davidson, Denise L. Mauzerall, Timothy D. Searchinger, Patrice Dumas, and Ye Shen.
2015. "Managing Nitrogen for Sustainable Development". Nature 528 (7580): 51–59. https://doi.org/10.1038/nature15743.
}
\examples{
\dontrun{
a <- readSource("Zhang2015", convert = FALSE)
}

}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Benjamin Leon Bodirsky
}
