% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAdjustExperGuesses.R
\name{toolAdjustExperGuesses}
\alias{toolAdjustExperGuesses}
\title{toolAdjustExperGuesses}
\usage{
toolAdjustExperGuesses(q)
}
\arguments{
\item{q}{magpie object with expert guesses loaded}
}
\value{
Adjusted expert guesses
}
\description{
Function that uses ssp2 as baseline scenario and adjusts ssp1,ssp3,ssp4, and ssp5 based on
pre-established assumptions about the relative development in the different scenarios.
}
\author{
Marcos Alves
}
