% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAggregateCell2Country.R
\name{toolAggregateCell2Country}
\alias{toolAggregateCell2Country}
\title{toolAggregateCell2Country}
\usage{
toolAggregateCell2Country(x, weight = NULL, ...)
}
\arguments{
\item{x}{cellular magpie object with coordinates}

\item{weight}{aggregation weight}

\item{...}{additional options forwarded to \code{toolCountryFill}}
}
\value{
return country ISO level data
}
\description{
Aggregate cellular data (with coordinate information) to countries and perform consistency checks
}
\author{
Jan Philipp Dietrich
}
