% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCell2isoCell.R
\name{toolCell2isoCell}
\alias{toolCell2isoCell}
\title{toolCell2isoCell}
\usage{
toolCell2isoCell(x, cells = "magpiecell")
}
\arguments{
\item{x}{magpie object on cellular level}

\item{cells}{switch between magpie cells (59199) and lpj cells (67420)}
}
\value{
return changed input data
}
\description{
Sets cell names to "iso country code"."cell number"
}
\author{
Kristine Karstens
}
