% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolClimateInputVersion.R
\name{toolClimateInputVersion}
\alias{toolClimateInputVersion}
\title{toolClimateInputVersion}
\usage{
toolClimateInputVersion(lpjmlVersion, climatetype)
}
\arguments{
\item{lpjmlVersion}{Add-ons (+*) for further version specification for LPJmL version}

\item{climatetype}{Switch between different climate scenarios}
}
\value{
configuration as list
}
\description{
Specify default settings for LPJmL climate input version and baseline settings
}
\author{
Kristine Karstens
}
