% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolConv2CountryByCelltype.R
\name{toolConv2CountryByCelltype}
\alias{toolConv2CountryByCelltype}
\title{toolConv2CountryByCelltype}
\usage{
toolConv2CountryByCelltype(x, cells)
}
\arguments{
\item{x}{magpie object on cellular level}

\item{cells}{switch between 59199 ("magpiecell") and 67420 ("lpjcell") cells}
}
\value{
return selected input data on ISO country level
}
\description{
Aggregates cellular data to ISO country level after conversion of cellular
data to a specific cell setup (this type is relevant as some settings,
such as "magpiecell" remove some cells and therby affect country sums)
}
\author{
Jan Philipp Dietrich
}
