% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCoord2Isocell.R
\name{toolCoord2Isocell}
\alias{toolCoord2Isocell}
\title{toolCoord2Isocell}
\usage{
toolCoord2Isocell(x, cells = "magpiecell", fillMissing = NULL)
}
\arguments{
\item{x}{object to be transformed from coordinates to (old) magpie isocell standard}

\item{cells}{Switch between "magpiecell" (59199) and "lpjcell" (67420)}

\item{fillMissing}{if NULL cells missing from the total 59199 are just being ignore. If set to a value
missing cells will be added with this value (e.g. all set to 0 if fillMissing is 0)}
}
\value{
magpie object with 59199 cells in isocell naming
}
\description{
Transforms an object with coordinate spatial data (on half-degree) to isocell (59199) standard
}
\author{
Kristine Karstens, Felicitas Beier, Jan Philipp Dietrich
}
