% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCoord2Isocoord.R
\name{toolCoord2Isocoord}
\alias{toolCoord2Isocoord}
\title{\if{html}{\out{<div class="sourceCode">}}\preformatted{  toolCoord2Isocoord
}\if{html}{\out{</div>}}}
\usage{
toolCoord2Isocoord(x)
}
\arguments{
\item{x}{object to be transformed from coordinates to iso-coordinate object}
}
\value{
magpie object with 67420 cells in x.y.iso naming
}
\description{
Transforms an object with coordinate spatial data (on half-degree)
to object with 67420 cells and coordinate and iso country information
}
\author{
Felicitas Beier
}
