% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolCountryFillBilateral.R
\name{toolCountryFillBilateral}
\alias{toolCountryFillBilateral}
\title{toolCountryFillBilateral}
\usage{
toolCountryFillBilateral(x, fill = NA)
}
\arguments{
\item{x}{input variable, a bilateral magclass object}

\item{fill}{fill value, default NA}
}
\description{
Fills bilateral iso-level magpie objects to 249 x 249 countries
}
