% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolExtrapolateFodder.R
\name{toolExtrapolateFodder}
\alias{toolExtrapolateFodder}
\title{\preformatted{  toolExtrapolateFodder
}}
\usage{
toolExtrapolateFodder(x, exyears = c(2004, 2009), average = 5)
}
\arguments{
\item{x}{input data}

\item{exyears}{two years}

\item{average}{the averaging_range in toolTimeInterpolate}
}
\value{
magpie object including extrapolated years
}
\description{
Extrapolate fodder data, based on two time steps (5-averages around this years)
}
\author{
Kristine Karstens
}
