% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFAOForestRelocate.R
\name{toolFAOForestRelocate}
\alias{toolFAOForestRelocate}
\title{toolFAOForestRelocate}
\usage{
toolFAOForestRelocate(luInit, forestArea, vegC)
}
\arguments{
\item{luInit}{uncorrected landuse initialisation data set}

\item{forestArea}{forest target values on country level}

\item{vegC}{vegetation carbon data used as reallocation weight}
}
\value{
List of magpie object with results on cellular level
}
\description{
Reallocates cellular forest information from LUH2 to better match FAO
forest information
}
\author{
Kristine Karstens, Felicitas Beier, Patrick v. Jeetze, Jan Philipp Dietrich
}
