% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolFixIEAdataForIndustrySubsectors.R
\name{toolFixIEAdataForIndustrySubsectors}
\alias{toolFixIEAdataForIndustrySubsectors}
\title{Apply adjustments to industry-related IEA data}
\usage{
toolFixIEAdataForIndustrySubsectors(data, fixing, threshold = 0.01)
}
\arguments{
\item{data}{MAgPIE object containing the IEA Energy Balances data}

\item{fixing}{temporary flag}

\item{threshold}{minimum share each industry subsector uses of each product.
Defaults to 1 \%.}
}
\value{
a MAgPIE object
}
\description{
This function prepares the industry-related IEA before mapping it to REMIND sectors.
There are three different types of adjustments done:
\enumerate{
\item replace coke oven and blast furnace outputs (\code{BLFURGS}, \code{OGASES}, \code{OVENCOKE},
\code{COKEOVGS}, \code{COALTAR}, \code{NONCRUDE}) by inputs
(required for dealing with energy flows from the steel sector to other sectors)
\item prepare industry-related time series
\item apply corrections to IEA data to cope with fragmentary time series
}
}
\details{
The corrections done by this function are rather rudimentary and crude. This
gets smoothed away in regional aggregation. But do not use the resulting
country-level data without additional scrutiny.

Use regional or global averages if IEA industry data lists energy use only as
"non-specified".
}
\author{
Michaja Pehl, Felix Schreyer
}
