% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolForestRelocate.R
\name{toolForestRelocate}
\alias{toolForestRelocate}
\title{toolForestRelocate}
\usage{
toolForestRelocate(lu, luCountry, natTarget, vegC)
}
\arguments{
\item{lu}{uncorrected landuse initialisation data set (cell level)}

\item{luCountry}{uncorrected landuse initialisation on country level}

\item{natTarget}{target natural land allocation on country level}

\item{vegC}{vegetation carbon data used as reallocation weight}
}
\value{
List of magpie object with results on cellular level
}
\description{
Reallocates cellular forest information from LUH2 to better match FAO
forest information
}
\author{
Kristine Karstens, Felicitas Beier, Patrick v. Jeetze, Jan Philipp Dietrich
}
