% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolHarmonize2Baseline.R
\name{toolHarmonize2Baseline}
\alias{toolHarmonize2Baseline}
\title{toolHarmonize2Baseline}
\usage{
toolHarmonize2Baseline(
  x,
  base,
  ref_year = "y2015",
  limited = TRUE,
  hard_cut = FALSE
)
}
\arguments{
\item{x}{magclass object that should be set on baseline}

\item{base}{magclass object for baseline}

\item{ref_year}{Reference year}

\item{limited}{if TRUE, for an underestimated baseline the signal is limited to the additive term rather than the multiplicative factor
if FALSE, x will be put on the baseline as a pure multiplicative factor}

\item{hard_cut}{Switch to TRUE for data that can not be harmonized, but have to be glued together}
}
\value{
the averaged data in magclass format
}
\description{
toolHarmonize2Baseline
}
\author{
Kristine Karstens
}
