% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSmooth.R
\name{toolSmooth}
\alias{toolSmooth}
\title{toolSmooth}
\usage{
toolSmooth(x, method = "spline")
}
\arguments{
\item{x}{magclass object that should be smoothed}

\item{method}{spline, average or more (See default argument for current default setting)}
}
\value{
smoothed data in magclass format
}
\description{
Smooth a time series using a given method and its default settings
}
\author{
Kristine Karstens
}
