% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSplitBiomass.R
\name{toolSplitBiomass}
\alias{toolSplitBiomass}
\title{Split Biomass into modern and traditional}
\usage{
toolSplitBiomass(
  x,
  gdppop,
  split = "biomass",
  into = c("biotrad", "biomod"),
  dim = 3.1,
  limits = c(12300, 18500)
)
}
\arguments{
\item{x}{MagPIE object including biomass data}

\item{gdppop}{MagPIE object with GDP/cap data}

\item{split}{character, name of item to split}

\item{into}{character vector of length two with the names of the split items}

\item{dim}{dimension of \code{x} with the item to split}

\item{limits}{numeric vector of length two with the corresponding GDP/cap
limits. The default values used to be 10k and 15k USD/cap converted from
2005 to 2017 dollars.}
}
\value{
MagPIE object including split biomass data
}
\description{
We assume that below a given GDP/cap level, all biomass is traditional and
above a higher limit, all biomass is modern with a linear transition between
the limits.
}
\author{
Robin Hasse
}
