% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolSum2Country.R
\name{toolSum2Country}
\alias{toolSum2Country}
\title{toolSum2Country}
\usage{
toolSum2Country(x)
}
\arguments{
\item{x}{magpie object on cellular level with countries in dim 1.1}
}
\value{
return selected input data on ISO country level
}
\description{
Efficient method to sum cellular data with country dimension as first
sub-dimension to country level
}
\author{
Jan Philipp Dietrich
}
