% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFacReqGLO.R
\name{calcFacReqGLO}
\alias{calcFacReqGLO}
\title{calcFacReqGLO}
\usage{
calcFacReqGLO()
}
\value{
MAgPIE object
}
\description{
This function calculates 2005 global factor requirement costs (05USDppp/tDM) using FAO databases
}
\examples{
\dontrun{
calcOutput("calcFacReqGLO")
}

}
\seealso{
\code{\link{calcOutput}},\code{\link{calcFactorIntensity}}
}
\author{
Edna J. Molina Bacca
}
