% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFactorCostsLivst.R
\name{calcFactorCostsLivst}
\alias{calcFactorCostsLivst}
\title{calcFactorCostsLivst}
\usage{
calcFactorCostsLivst(datasource = "USDA", otherLivst = FALSE)
}
\arguments{
\item{datasource}{only source available is "USDA" (calculates factor costs by applying factor cost share from USDA
to VoP from FAO)}

\item{otherLivst}{boolean: should FAO livestock categories that can't be matched to MAgPIE categories (i.e. beeswax,
wool, silkworms, and honey) be reported as "livst_other"?}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates factor costs for livestock production in mio. US$MER05
}
\examples{
\dontrun{
calcOutput("FactorCostsLivst")
}
}
\author{
Debbora Leip
}
