% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFallowLand.R
\name{calcFallowLand}
\alias{calcFallowLand}
\title{calcFallowLand}
\usage{
calcFallowLand(cellular = TRUE)
}
\arguments{
\item{cellular}{TRUE for cellular outputs.}
}
\value{
MAgPIE object containing fallow land in Mha
}
\description{
Calculates fallow land on grid cell level,
based on physical cropland extend and harvested area output
of LandInG data.
The formula
"fallow land are = max( physical cropland area - harvested cropland area, 0)"
is used.
Due to multiple cropping, harvested cropland area can be greater than non-fallow land area
and even greater than physical cropland area.
Thus, the results can only be considered a rough estimate of fallow land area.
}
\examples{
\dontrun{
calcOutput("FallowLand")
}
}
\seealso{
\code{\link{readLandInG}}
}
\author{
David Hoetten, Felicitas Beier
}
