% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFeedBasketsPast.R
\name{calcFeedBasketsPast}
\alias{calcFeedBasketsPast}
\title{calcFeedBasketsPast}
\usage{
calcFeedBasketsPast(non_eaten_food = TRUE)
}
\arguments{
\item{non_eaten_food}{if TRUE, non-eaten food is included in feed baskets,
if not it is excluded.}
}
\value{
Historical feed baskets and corresponding weights
as a list of two MAgPIE objects
}
\description{
Calculate historical feed baskets
based on output of MAgPIE_FEED model
as DM feed biomass (different types of feed)
needed per DM livestock products
}
\examples{
\dontrun{
calcOutput("FeedBasketsPast")
}
}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}, \code{\link[=readFeedModel]{readFeedModel()}}
}
\author{
Isabelle Weindl, Benjamin Bodirsky
}
