% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPppp.R
\name{calcGDPpppFuture}
\alias{calcGDPpppFuture}
\title{calcGDPpppFuture}
\usage{
calcGDPpppFuture(GDPpppFuture = NULL)
}
\arguments{
\item{GDPpppFuture}{GDPppp future data source}
}
\value{
GDP PPP(ICP11) in million USD05 equivalents
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is defunct. Use \code{\link[mrdrivers:calcGDPFuture]{mrdrivers::calcGDPFuture()}} instead.
}
\seealso{
\code{\link[mrdrivers:calcGDPFuture]{mrdrivers::calcGDPFuture()}}
}
