% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGDPppp.R
\name{calcGDPpppPast}
\alias{calcGDPpppPast}
\title{calcGDPpppPast}
\usage{
calcGDPpppPast(GDPpppPast = "WDI-MI")
}
\arguments{
\item{GDPpppPast}{GDPppp future data source}
}
\value{
GDP PPP(ICP11) in million USD05 equivalents
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function is defunct. Use \code{\link[mrdrivers:calcGDPPast]{mrdrivers::calcGDPPast()}} instead.
}
\seealso{
\code{\link[mrdrivers:calcGDPPast]{mrdrivers::calcGDPPast()}}
}
