% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockNRF.R
\name{calcGrowingStockNRF}
\alias{calcGrowingStockNRF}
\title{calcGrowingStockNRF}
\usage{
calcGrowingStockNRF()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks from FAO data for naturally regenerating forests
i.e. primary forests and secondary forests.
}
\examples{
\dontrun{
calcOutput("GrowingStockNRF", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra
}
