% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcGrowingStockPlantations.R
\name{calcGrowingStockPlantations}
\alias{calcGrowingStockPlantations}
\title{calcGrowingStockPlantations}
\usage{
calcGrowingStockPlantations()
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
Calculates the growing stocks on FAO data.
}
\examples{
\dontrun{
calcOutput("GrowingStockPlantations", aggregate = TRUE)
}

}
\author{
Abhijeet Mishra
}
