% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcIOEdgeBuildings.R
\name{calcIOEdgeBuildings}
\alias{calcIOEdgeBuildings}
\title{calcIOEdgeBuildings}
\usage{
calcIOEdgeBuildings(
  subtype = c("output_EDGE", "output_EDGE_buildings"),
  ieaVersion = c("default", "latest")
)
}
\arguments{
\item{subtype}{Data subtype. See default argument for possible values.}

\item{ieaVersion}{Release version of IEA data, either 'default'
(vetted and used in REMIND) or 'latest'.}
}
\value{
IEA data as MAgPIE object aggregated to country level
}
\description{
Calculates buildings-related energy flows from the IEA energy balances.
'output_EDGE_buildings' is a key input to EDGE-Buildings providing the
historic final energy demand from buildings. 'output_EDGE' does the same for
buildings and industry together.
}
\examples{
\dontrun{
a <- calcOutput("IOEdgeBuildings", subtype = "output_EDGE_buildings")
}

}
\seealso{
\code{\link{calcOutput}}
}
\author{
Pascal Sauer, Anastasis Giannousakis, Robin Hasse
}
