% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcNitrogenFixationFreeliving.R
\name{calcNitrogenFixationFreeliving}
\alias{calcNitrogenFixationFreeliving}
\title{calcNitrogenFixationPast}
\usage{
calcNitrogenFixationFreeliving(cellular = FALSE)
}
\arguments{
\item{cellular}{cellular estimates optional}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
calculates fixation from freeliving bacteria and from nitrogen-fixing crops
}
\examples{
\dontrun{
calcOutput("calcNitrogenFixationFreeliving")
}
}
\seealso{
\code{\link[=calcNitrogenFixationPast]{calcNitrogenFixationPast()}}
}
\author{
Benjamin Leon Bodirsky
}
