% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcProdSysRatioPast.R
\name{calcProdSysRatioPast}
\alias{calcProdSysRatioPast}
\title{Calculate historical distribution of livestock production across
different systems based on output of MAgPIE_FEED model}
\usage{
calcProdSysRatioPast(faoVersion = "join2010", yearly = FALSE)
}
\arguments{
\item{faoVersion}{which version of FAO food balances to use in mass balance}

\item{yearly}{whether to calculate yearly data or only magpie 5year timesteps}
}
\value{
Historical distribution of livestock production across
different systems and corresponding weights as a list of two MAgPIE objects
}
\description{
Calculate historical distribution of livestock production across
different systems based on output of MAgPIE_FEED model
}
\examples{
\dontrun{
calcOutput("ProdSysRatioPast")
}
}
\seealso{
\code{\link[madrat:calcOutput]{madrat::calcOutput()}}, \code{\link[=readFeedModel]{readFeedModel()}}
}
\author{
Isabelle Weindl
}
