% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcRockNWeathering.R
\name{calcRockNWeathering}
\alias{calcRockNWeathering}
\title{calcRockNWeathering}
\usage{
calcRockNWeathering()
}
\value{
MAgPIE object of amount of N (Mt)
}
\description{
calculates amount of yearly N from rock weathering by country or global total,
disaggregated by land use type (6 classes from LanduseInitialisation)
}
\author{
David M Chen
}
