% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convertILOSTAT.R
\name{convertILOSTAT}
\alias{convertILOSTAT}
\title{convertILOSTAT}
\usage{
convertILOSTAT(x)
}
\arguments{
\item{x}{unconverted magpie object from read-script}
}
\value{
Data as MAgPIE object with common country list
}
\description{
transforms currencies where applicable, and fills missing countries in ILOSTAT data with 0
}
\examples{
\dontrun{
  a <- readSource("ILOSTAT", subtype = "EmplByActivityModelled", convert = TRUE)
}
}
\author{
Debbora Leip
}
