% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correctLPJmLInputs.R
\name{correctLPJmLInputs}
\alias{correctLPJmLInputs}
\title{\preformatted{  correctLPJmLInputs
}}
\usage{
correctLPJmLInputs(x)
}
\arguments{
\item{x}{magpie object provided by the read function}
}
\description{
correct LPJmLInputs content (dummy function)
}
\examples{
\dontrun{
readSource("LPJmLInputs", convert = "onlycorrect")
}

}
\author{
Felicitas Beier
}
