% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrcommons-package.R
\docType{package}
\name{mrcommons-package}
\alias{mrcommons-package}
\alias{mrcommons}
\title{MadRat commons Input Data Library}
\description{
Provides useful functions and a common structure to all the input data required to run models like MAgPIE and REMIND of model input data.
}
\author{
Maintainer: Jan Philipp Dietrich \href{mailto:dietrich@pik-potsdam.de}{dietrich@pik-potsdam.de}
}
