% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCMIP7_CEDS.R
\name{readCMIP7_CEDS}
\alias{readCMIP7_CEDS}
\title{readCMIP7_CEDS}
\usage{
readCMIP7_CEDS()
}
\value{
MAgPIE object
}
\description{
reads in harmonized emission data used for CMIP7 from three files.
country_file for energy related emissions on country level and global bunker
emissions, global_file for F-Gases and AFOLU on global level,
cmip7-historical-ghg-eq for GHG totals.
Source: https://zenodo.org/records/15059443
}
\author{
Pascal Weigmann
}
