% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEDGARghg.R
\name{readEDGARghg}
\alias{readEDGARghg}
\title{readEDGAR_GHG}
\usage{
readEDGARghg()
}
\value{
magpie objects of EDGAR historical emissions (MtCO2eq)
}
\description{
read EDGAR GHG emission data
}
\author{
Pascal Weigmann
}
