% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readEurostat.R
\name{readEurostat}
\alias{readEurostat}
\title{Read Eurostat historical emissions (env_air_gge)}
\usage{
readEurostat(subtype = "emissions")
}
\arguments{
\item{subtype}{'emissions' for original Eurostat emissions split,
'MACCemi' for MACC historical emissions, or 'sectorEmi' for sector specific
emissions, or 'latest' for most up-to-date data}
}
\value{
magpie object of Eurostat historical emissions (MtCO2)
}
\description{
Read-in Eurostat historical emissions csv files as magclass object
}
\examples{
\dontrun{
a <- readSource(type = "Eurostat", subtype = "emissions")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Renato Rodrigues
}
