% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readExpertGuessSSPLivestockProduction.R
\name{readExpertGuessSSPLivestockProduction}
\alias{readExpertGuessSSPLivestockProduction}
\title{readExpertGuessSSPLivestockProduction}
\usage{
readExpertGuessSSPLivestockProduction(subtype)
}
\arguments{
\item{subtype}{: Available subtypes are: ssp1 to 5 Data for the SSP Scenario}
}
\value{
magpie object containing the expert guesses
}
\description{
{Read the Expert Guesses for future Livestock Production for the SSP Scenarios
}
}
\examples{
\dontrun{
a <- readSource("ExpertGuessSSPLivestockProduction","ssp1")
}
}
\seealso{
\code{\link[madrat:readSource]{madrat::readSource()}}
}
\author{
Stephen Wirth
}
