% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readFRA2020.R
\name{readFRA2020}
\alias{readFRA2020}
\title{Read FRA2020}
\usage{
readFRA2020(subtype)
}
\arguments{
\item{subtype}{data subtype. Available subtypes: "forest_area","growing_stock","biomass_stock","carbon_stock","management","disturbance","forest_fire"}
}
\value{
Magpie object of the FRA 2020 data
}
\description{
Read-in an FRA data from 2020 (forest resource assessment).
}
\examples{

\dontrun{ a <- readSource("FRA2020","growing_stock")
}

}
\seealso{
\code{\link{readSource}}
}
\author{
Abhijeet Mishra
}
