% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGGCMICropCalendar.R
\name{readGGCMICropCalendar}
\alias{readGGCMICropCalendar}
\title{readGGCMICropCalendar}
\usage{
readGGCMICropCalendar(subtype = "fraction_of_harvested_area")
}
\arguments{
\item{subtype}{variable or vector of variables to read from the crop calendar set. Options:
("planting_day","maturity_day","fraction_of_harvested_area")}
}
\value{
MAgPIE object with the requested data
}
\description{
Reads in GGCMI fraction of Harvested Area masks for rice 1 and rice 2
(other crops available too, see path in download fucntion), or other variables available in
the GGCMI crop calendar.
}
\author{
David M Chen, Edna Molina Bacca
}
