% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readGGCMICropCalendar.R
\name{readGGCMICropCalendar}
\alias{readGGCMICropCalendar}
\title{readGGCMICropCalendar}
\usage{
readGGCMICropCalendar(subtype)
}
\arguments{
\item{subtype}{"cal" for the basic calendar "wheat_areas" for swh and wwh growing areas, "rice_areas" for}
}
\value{
A MAgPIE object, day of the year
}
\description{
Reads in planting, maturity, and harvest date based on GGCMI hybrid product of crop calendars. Note that 0 means crop is not harvested in a given grid cell.
}
\author{
David Chen
}
