% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readICP2017.R
\name{readICP2017}
\alias{readICP2017}
\title{readICP2017}
\usage{
readICP2017(subtype = "per_cap_expMER")
}
\arguments{
\item{subtype}{data subtype needed. Either "priceLevel", or "per_cap_exp"}
}
\value{
magpie object of relative price levels (world = 100) or per capita expenditure (USD17 MER)
}
\description{
Reads data of World Bank ICP round, downloaded from here: https://databank.worldbank.org/source/icp-2017
Reads
}
\examples{
\dontrun{
a <- readSource("ICP2017", "per_cap_exp")
}

}
\author{
David M Chen
}
