% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toolAWMSScenarioCreation.R
\name{toolAWMSScenarioCreation}
\alias{toolAWMSScenarioCreation}
\title{toolAWMSScenarioCreation}
\usage{
toolAWMSScenarioCreation(name, start_year, categories, values, out)
}
\arguments{
\item{name}{Name of the scenario}

\item{start_year}{Year were prediction starts}

\item{categories}{share of manure managed in different animal waste management systems}

\item{values}{target values}

\item{out}{contains historical data}
}
\value{
List of magpie objects with results on country level, weight on country level, unit and description.
}
\description{
tool function to calculated the share of manure managed in different animal waste management systems in confinements.
}
\seealso{
\code{\link{calcAWMSconfShr}}
}
\author{
Edna J. Molina Bacca
}
