% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calcFAOForestRelocate.R
\name{toolFAOForestRelocate}
\alias{toolFAOForestRelocate}
\title{toolFAOForestRelocate}
\usage{
toolFAOForestRelocate(luInit, luTargetCountry, vegC)
}
\arguments{
\item{selectyears}{default on "past"}

\item{cells}{if cellular is TRUE: "magpiecell" for 59199 cells or "lpjcell" for 67420 cells}
}
\value{
List of magpie object with results on cellular level, weight on cellular level, unit and description.
}
\description{
Calculates the cellular MAgPIE forest and other land area correction based on FAO forestry data
and LUH2v2.
}
\author{
Kristine Karstens, Felicitas Beier, Patrick v. Jeetze, Jan Philipp Dietrich
}
